% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust_helpers.R
\name{calculate_weighted_dist}
\alias{calculate_weighted_dist}
\title{calculate_weighted_dist: Calculate weighted distance matrices}
\usage{
calculate_weighted_dist(scaled_dists = NULL, ws = NULL)
}
\arguments{
\item{scaled_dists}{a list with two matrices: scaled expression and spatial distance matrices}

\item{ws}{vector of spatial weights (0-1) to apply to spatial distances}
}
\value{
a list of weighted distance matrices, one for each ws value
}
\description{
Combine scaled expression and spatial distance matrices using user-specified weights
}
\keyword{internal}
