% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_conversion.R
\name{can_convert_to_STlist}
\alias{can_convert_to_STlist}
\title{Check if object can be converted to STlist}
\usage{
can_convert_to_STlist(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical TRUE if conversion is possible
}
\description{
Tests whether an object can be converted to STlist format
}
\keyword{internal}
