% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_geary_C.R
\name{cell_geary_C}
\alias{cell_geary_C}
\title{cell_geary_C}
\usage{
cell_geary_C(x = NULL, cells = NULL, subj = NULL)
}
\arguments{
\item{x, }{an STList with deconvoluted cell data.}

\item{cells, }{a vector of cell names present in the deconvolution matrix.}

\item{subj, }{an integer indicating the spatial array for which the statistic
will be calculated.}
}
\value{
x, a STList including the values corresponding to Geary's C for each
cell in cells.
}
\description{
Calculates Geary's C from cell deconvolution scores in ST data.
}
\details{
This function takes an STList and a vector with cell names (from xCell, for
example) and returns Geary's C for each element of the vector.
}
