% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_getis_Gi.R
\name{cell_getis_Gi}
\alias{cell_getis_Gi}
\title{cell_getis_Gi}
\usage{
cell_getis_Gi(x = NULL, cells = NULL, subj = NULL, method = "xcell")
}
\arguments{
\item{x, }{an STList with deconvoluted cell data.}

\item{cells, }{a vector of cell names present in the deconvolution matrix.}

\item{subj, }{an integer indicating the spatial array for which the statistic
will be calculated.}

\item{method, }{the deconvolutionn method to used estimate the cell scores.}
}
\value{
x, a STList including the values corresponding to Getis-Ord Gi for each
cell in cells.
}
\description{
Calculates Getis-Ord Gi C from cell deconvolution scores in ST data.
}
\details{
This function takes an STList and a vector with cell names (from xCell, for
example) and returns Getis-Ord Gi for each element of the vector.
}
