% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_moran_I.R
\name{cell_moran_I}
\alias{cell_moran_I}
\title{cell_moran_I}
\usage{
cell_moran_I(x = NULL, cells = NULL, subj = NULL)
}
\arguments{
\item{x, }{an STList with deconvoluted cell data.}

\item{cells, }{a vector of cell names present in the deconvolution matrix.}

\item{subj, }{an integer indicating the spatial array for which the statistic
will be calculated.}
}
\value{
x, a STList including the values corresponding to Moran's I for each
cell in cells.
}
\description{
Calculates Moran's I from cell deconvolution scores in ST data.
}
\details{
This function takes an STList and a vector with cell names (from xCell, for
example) and returns Moran's I for each element of the vector.
}
