% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_purity.R
\name{cluster_purity}
\alias{cluster_purity}
\title{cluster_purity}
\usage{
cluster_purity(x = NULL)
}
\arguments{
\item{x, }{an STList.}

\item{who, }{an integer indicating the spatial array to be analyzed.}
}
\description{
Perform model-based clustering (using BIC) to define stromal/tumor
clusters from ESTIMATE purity scores.
}
\details{
yada yada yada
}
