% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_parse}
\alias{color_parse}
\title{color_parse: Creates a color palette}
\usage{
color_parse(color_pal = NULL, n_cats = NULL)
}
\arguments{
\item{color_pal}{A name of a Khroma or RColorBrewer. Alternatively, a vector with
colors of lenght equal or larger than number of categories.}

\item{n_cats}{The number of colors to produce. If NULL, assumes 5 colors.}
}
\value{
cat_cols A color palette
}
\description{
Uses Khroma or RColorBrewer to return the colors of a palette name.
}
\details{
This function takes a character string and uses either khroma or RColoBrewer to
create a color palette. The function first looks in khroma, then RColoBrewer. In
other words, khroma colors have priority.
}
