% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_helpers.R
\name{count_cores}
\alias{count_cores}
\title{count_cores: Determine optimal number of cores for parallelization}
\usage{
count_cores(n)
}
\arguments{
\item{n}{an integer representing the number of units to process}
}
\value{
the number of cores to use (max = half of available cores)
}
\description{
Determines the number of cores to use for parallelization
}
\keyword{internal}
