% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_listw_from_knn}
\alias{create_listw_from_knn}
\title{create_listw_from_knn}
\usage{
create_listw_from_knn(x = NULL, ks = NULL)
}
\arguments{
\item{x}{an STlist}
}
\description{
create_listw_from_knn
}
