% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_input.R
\name{detect_input}
\alias{detect_input}
\title{detect_input: Determine what is being provided to STList}
\usage{
detect_input(rnacounts = NULL, spotcoords = NULL, samples = NULL)
}
\arguments{
\item{rnacounts, }{the file/directory with counts provided to STList.}

\item{spotcoords, }{the file with coordinates  provided to STList.}

\item{samples, }{the metadata or sample names provided to STList.}
}
\value{
inputtype, a list containing file types of input arguments.
}
\description{
Detects the type of input being provided to the fucntion STList.
}
\details{
This function detects what input is being provided to the STList() function. It
also detects the delimiter of the file when relevant. NOTE that the function does
minimum checking on the contents of file, limited mostly to detect if file is csv or tsv,
or if Visium files are available. Checks are performed on the first element only,
and thus other elements could not comply with the format.
}
