% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\name{dim,STlist-method}
\alias{dim,STlist-method}
\title{dim: Prints the dimensions of count arrays within an STList object.}
\usage{
\S4method{dim}{STlist}(x)
}
\arguments{
\item{x}{an STList object to show summary from.}
}
\description{
Returns the number of genes and spots for each array within an STList object
}
\details{
This function takes an STList and prints the number of genes (rows) and spots (columns) of
each spatial array within that object.
}
