% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_generic.R
\name{dispatch_ingest.source_list}
\alias{dispatch_ingest.source_list}
\title{Ingest List of DataFrames}
\usage{
\method{dispatch_ingest}{source_list}(source)
}
\arguments{
\item{source}{InputSource object of type 'list'}
}
\value{
list containing counts list and coords list (one per sample)
}
\description{
Handles named list of dataframes with counts and coordinates
}
\details{
Processes named list inputs where each element contains gene expression
dataframes. Sorts and aligns count and coordinate dataframes.
}
