% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_visium.R
\name{dispatch_ingest.source_visium}
\alias{dispatch_ingest.source_visium}
\title{Ingest Visium Data}
\usage{
\method{dispatch_ingest}{source_visium}(source)
}
\arguments{
\item{source}{InputSource object of type 'visium'}
}
\description{
Ingest Visium Data
}
