% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_helpers.R
\name{expandSparse}
\alias{expandSparse}
\title{expandSparse: Convert sparse matrix to dense data frame}
\usage{
expandSparse(sparsedMatrix)
}
\arguments{
\item{sparsedMatrix}{a sparse matrix to convert}
}
\value{
a dense data frame
}
\description{
Expands a sparse matrix to a dense data frame
}
\keyword{internal}
