% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_interpolation.R
\name{gene_interpolation}
\alias{gene_interpolation}
\title{gene_interpolation: Spatial interpolation of gene expression}
\usage{
gene_interpolation(
  x = NULL,
  genes = "top",
  top_n = 10,
  samples = NULL,
  ngrid = 10000,
  REML = F,
  cores = NULL
)
}
\arguments{
\item{x}{an STlist with transformed RNA counts}

\item{genes}{a vector of gene names or 'top'. If 'top' (default), interpolation of
the 10 genes (\code{top_n} default) with highest standard deviation in each ST sample
is estimated.}

\item{top_n}{an integer indicating how many top genes to perform interpolation.
Default is 10.}

\item{samples}{the spatial samples for which interpolations will be performed.
If NULL (Default), all samples are interpolated.}

\item{ngrid}{an integer indicating the number of point to predict. Default is 10000,
resulting in a grid of 100 x 100 points. Larger numbers provide more resolution,
but computing time is longer.}

\item{REML}{wheter to use restricted maximum likelihood to estimate spatial covariance
parameters. Default is FALSE. Setting to TRUE might speed up interpolation, but some
issue have been found.}

\item{cores}{integer indicating the number of cores to use during parallelization.
If NULL, the function uses half of the available cores at a maximum. The parallelization
uses \code{parallel::mclapply} and works only in Unix systems.}
}
\value{
x a STlist including spatial interpolations.
}
\description{
Performs spatial interpolation ("kriging") of transformed gene counts
}
\details{
This function takes an STlist and a vector of gene names and generates spatial
interpolation of gene expression values via "kriging". If genes='top', then
the 10 genes (default) with the highest standard deviation for each ST sample
are interpolated. The resulting interpolations can be visualized via the
\code{STplot_interpolation} function
}
\examples{
# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- system.file("extdata", 'melanoma_thrane', package="spatialGE")
count_files <- list.files(data_files, full.names=TRUE, pattern='counts')
coord_files <- list.files(data_files, full.names=TRUE, pattern='mapping')
clin_file <- list.files(data_files, full.names=TRUE, pattern='clinical')
melanoma <- STlist(rnacounts=count_files[c(1,2)], spotcoords=coord_files[c(1,2)], samples=clin_file) # Only first two samples
melanoma <- transform_data(melanoma)
melanoma <- gene_interpolation(melanoma, genes=c('MLANA', 'COL1A1'), samples='ST_mel1_rep2')
kp = STplot_interpolation(melanoma, genes=c('MLANA', 'COL1A1'))
ggpubr::ggarrange(plotlist=kp)

}
