% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_moran_I.R
\name{gene_moran_I}
\alias{gene_moran_I}
\title{gene_moran_I}
\usage{
gene_moran_I(x = NULL, genes = NULL, subj = NULL)
}
\arguments{
\item{x, }{an STList with normalized gene counts.}

\item{genes, }{a vector with gene names in the normalized count matrix.}

\item{subj, }{the index of the spatial array for which the statistic
will be calculated.}
}
\value{
x, a STList including the values corresponding to Moran's I for each
gene in genes.
}
\description{
Calculates Moran's I from ST data.
}
\details{
This function takes an STList and a vector with HUGO gene names and returns
Morans' I for each element of the vector.
}
