% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stat_plot.R
\name{get_gene_stats}
\alias{get_gene_stats}
\title{get_gene_stats}
\usage{
get_gene_stats(x = NULL, gene = NULL, clinvar_vals = NULL)
}
\arguments{
\item{x}{an STList}

\item{gene}{a gene name}

\item{clinvar_vals}{a data frame to populate with spatial statistics}
}
\description{
get_gene_stats
}
