% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust_helpers.R
\name{get_hier_clusters_dtc}
\alias{get_hier_clusters_dtc}
\title{get_hier_clusters_dtc: Hierarchical clustering with DynamicTreeCut}
\usage{
get_hier_clusters_dtc(
  weighted_dists = NULL,
  ws = NULL,
  deepSplit = NULL,
  linkage = NULL
)
}
\arguments{
\item{weighted_dists}{a list of distance matrices (NOT dist objects) for each spatial weight}

\item{ws}{a vector with spatial weights}

\item{deepSplit}{a logical or integer (1-4) for cutreeDynamic deepSplit parameter}

\item{linkage}{a string with the linkage method for hclust (e.g., 'ward.D2')}
}
\value{
a list of data frames with spot/cell cluster assignments for each weight
}
\description{
Perform hierarchical clustering followed by DynamicTreeCut for adaptive cluster detection
}
