% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust_helpers.R
\name{get_hier_clusters_ks}
\alias{get_hier_clusters_ks}
\title{get_hier_clusters_ks: Hierarchical clustering with fixed k values}
\usage{
get_hier_clusters_ks(
  weighted_dists = NULL,
  ws = NULL,
  ks = NULL,
  linkage = NULL
)
}
\arguments{
\item{weighted_dists}{a list of distance matrices (NOT dist objects) for each spatial weight}

\item{ws}{a vector with spatial weights}

\item{ks}{a vector with k values for cluster detection}

\item{linkage}{a string with the linkage method for hclust (e.g., 'ward.D2')}
}
\value{
a list of data frames with spot/cell cluster assignments for each weight
}
\description{
Perform hierarchical clustering followed by cutree for fixed number of clusters
}
\keyword{internal}
