% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_visium.R
\name{importVisium}
\alias{importVisium}
\title{importVisium: Reads Visium outputs and produce an STList.}
\usage{
importVisium(visiumfp = NULL, filterMT = T, savefiles = T, stlist = T)
}
\arguments{
\item{visiumfp, }{the file path to an 'outs' folder from a Visium run.}

\item{filterMT, }{logical, whether or not to filter mtDNA genes. Filters out gene names
beginning with 'MT-'. Defaults to TRUE.}

\item{savefiles, }{logical, whether or not to save the counts and coordinate files.
The files are saved as 'visium_counts.txt' and visium_coords.txt'. Defaults to TRUE.}

\item{stlist, }{logical, whther or not return an STList.}
}
\value{
x, an STList with the Visium counts and coordinates.
}
\description{
Reads files in the output folder of a Visium run, and returns an
STList for downstream analysis with spatialGE.
}
\details{
The function takes as an argument the path to an 'outs'  folder of a Visium run.
It reads the data and converts it into an STList, which can be used in downstrean
analysis with spatialGE. Optionally, the function can also output the count and
coordinates files.
}
