% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{krige_p}
\alias{krige_p}
\title{krige_p}
\usage{
krige_p(
  data_f = NULL,
  mask = NULL,
  color_pal = "YlOrBr",
  leg_name = "",
  title_name = "",
  minvalue = minvalue,
  maxvalue = maxvalue,
  visium = T
)
}
\arguments{
\item{data_f, }{a data with three columns: x coordinates, y coordinates, and
the kriging prediction values to be plotted.}

\item{mask, }{an object of class SpatialPolygons containing a large polygon
encasing all the predicted grid, and a smaller polygon drawing the concave hull
of the tissue shape.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{leg_name, }{a short name for the legend title.}

\item{title_name, }{a short name for the plot title.}

\item{minvalue, }{the minimum value of gene expression or cell score. Used for
standardization.}

\item{maxvalue, }{the maximum value of gene expression or cell score. Used for
standardization.}

\item{visium, }{whether or not to reverse axes for Visium slides.}
}
\value{
, a ggplot object.
}
\description{
Creates a kriging plot from ST data.
}
\details{
Function to produce a "kriging plot" from a data frame with three columns:
x coordinates, y coordinates, and predicted kriging values. The data frame has
column names 'x_pos', 'y_pos', and 'krige'. The function also takes a
SpatialPolygons object to mask the predicted grid to the area of the tissue.
It also takes a color palette name from the 'khroma' package. Finally, it
takes a name for the color legend title.
}
