% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{krige_p_pvals}
\alias{krige_p_pvals}
\title{krige_p_pvals}
\usage{
krige_p_pvals(
  data_f = NULL,
  mask = NULL,
  color_pal = "YlOrBr",
  leg_name = "",
  title_name = "",
  x = NULL,
  plot_who = NULL,
  cell = NULL,
  minvalue = minvalue,
  maxvalue = maxvalue,
  visium = T
)
}
\arguments{
\item{data_f, }{a data with three columns: x coordinates, y coordinates, and
the kriging prediction values to be plotted.}

\item{mask, }{an object of class SpatialPolygons containing a large polygon
encasing all the predicted grid, and a smaller polygon drawing the concave hull
of the tissue shape.}

\item{color_pal, }{a scheme from 'khroma'. Defaul is 'YlOrBr'}

\item{leg_name, }{a short name for the legend title.}

\item{title_name, }{a short name for the plot title.}

\item{x, }{an STList from where coordinates will be taken.}

\item{plot_who, }{an integer indicating the spatial array to be plotted.}

\item{cell, }{a cell name from a deconvoluted matrix.}

\item{minvalue, }{the minimum value of gene expression or cell score. Used for
standardization.}

\item{maxvalue, }{the maximum value of gene expression or cell score. Used for
standardization.}

\item{visium, }{whether or not to reverse axes for Visium slides.}
}
\value{
, a ggplot object.
}
\description{
Creates a kriging plot from ST data with xCell pvalues.
}
\details{
Function to produce a "kriging plot" from a data frame with three columns:
x coordinates, y coordinates, and predicted kriging values. The data frame has
column names 'x_pos', 'y_pos', and 'krige'. The function also takes a
SpatialPolygons object to mask the predicted grid to the area of the tissue.
It also takes a color palette name from the 'khroma' package. Finally, it
takes a name for the color legend title. Finally, the function overlays dots
for spots in which significant p-values (<0.05) were observed for a cell type.
}
