% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_images}
\alias{load_images}
\title{load_images: Place tissue images within STList}
\usage{
load_images(x = NULL, images = NULL)
}
\arguments{
\item{x}{an STList with transformed RNA counts.}

\item{images}{a string indicating a folder to load images from}
}
\value{
an STList with images
}
\description{
Loads the images from tissues to the appropriate STList slot.
}
\details{
This function looks for \code{.PNG} files within a folder matching the sample names
in an existing STList. Then, loads the images to the STList which can be used
for plotting along with quilt plots.
}
