% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTransform.R
\name{log_transf}
\alias{log_transf}
\title{log_transf: Log-transformation of spatial arrays}
\usage{
log_transf(x = NULL, scale_f = scale_f)
}
\arguments{
\item{x}{an STList with raw count matrices.}

\item{scale_f}{integer, the scaling factor to use.}
}
\value{
x an updated STList with transformed counts.
}
\description{
Applies the natural logarithm log(x+1) to counts after library
size normalization and multiplication by scaling factor (by default 10^4).
The function produces similar results to Seurat's LogNormalize.
}
\details{
The function works on parallel using "forking" (not in Windows OS).
}
