% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_utils.R
\name{new_input_source}
\alias{new_input_source}
\title{Constructor for InputSource}
\usage{
new_input_source(
  type,
  format = NULL,
  rna = NULL,
  coords = NULL,
  samples = NULL
)
}
\arguments{
\item{type}{character string of input type (visium, xenium, generic, seurat, list)}

\item{format}{character string detailed format (h5, mex, csv, etc)}

\item{rna}{object or path for counts}

\item{coords}{object or path for coordinates}

\item{samples}{object or path for samples metadata}
}
\value{
list with class attribute set
}
\description{
Constructor for InputSource
}
