% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_helpers.R
\name{non_spatial_de}
\alias{non_spatial_de}
\title{non_spatial_de: Run non-spatial linear models for gene x cluster combinations}
\usage{
non_spatial_de(expr_data = NULL, combo = NULL, pairwise = NULL)
}
\arguments{
\item{expr_data}{a data frame (rows=spots/cells) with normalized expression data
as well as coordinates (x,y), and a group variable for random effects.}

\item{combo}{a data frame with columns (samplename, meta1, meta2, gene) containing
all combinations of gene x cluster to test.}

\item{pairwise}{whether to perform pairwise tests (TRUE) or reference-based (FALSE)}
}
\value{
a list containing non-spatial model results with p-values, log-fold changes,
and metadata
}
\description{
Runs non-spatial linear models (or t-tests/Wilcoxon) for testing
differentially expressed genes between groups of spots/cells
}
\keyword{internal}
