% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_SThet.R
\name{old_SThet}
\alias{old_SThet}
\title{SThet: Computes global spatial autocorrelation statistics on gene expression}
\usage{
old_SThet(
  x = NULL,
  genes = NULL,
  samples = NULL,
  method = "moran",
  overwrite = T
)
}
\arguments{
\item{x}{an STlist}

\item{genes}{a vector of gene names to compute statistics}

\item{samples}{the samples to compute statistics}

\item{method}{The spatial statistic(s) to estimate. It can be set to 'moran',
'geary' or both. Default is 'moran'}

\item{overwrite}{logical indicating if previous statistics should be overwritten.
Default to TRUE (overwrite)}
}
\value{
an STlist containing spatial statistics
}
\description{
Computes the global spatial autocorrelation statistics Moran's I and/or
Geary's C for a set of genes
}
\details{
The function computes global spatial autocorrelation statistics (Moran's I and/or
Geary's C) for the requested genes and samples. Then computation uses the
package \code{spdep}. The calculated statistics are stored in the STlist, which can
be accessed with the \code{get_gene_meta} function.
}
\examples{
# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- list.files(system.file("extdata", package="spatialGE"), recursive=T, full.names=T)
count_files <- grep("counts", data_files, value=T)
coord_files <- grep("mapping", data_files, value=T)
clin_file <- grep("thrane_clinical", data_files, value=T)
melanoma <- STlist(rnacounts=count_files[c(1,2)], spotcoords=coord_files[c(1,2)], samples=clin_file) # Only first two samples
melanoma <- transform_data(melanoma, method='log')
melanoma <- SThet(melanoma, genes=c('MLANA', 'TP53'), method='moran')
get_gene_meta(melanoma, sthet_only=T)

}
