% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_krige.R
\name{plot_cell_krige}
\alias{plot_cell_krige}
\title{plot_cell_krige}
\usage{
plot_cell_krige(
  x = NULL,
  cells = NULL,
  krige_type = "ord",
  plot_who = NULL,
  color_pal = "YlOrBr",
  saveplot = F,
  pvalues = F
)
}
\arguments{
\item{x, }{an STList with kriging objects for the cells selected.}

\item{cells, }{a vector of cell names within a deconvolution matrix (one or
several) to plot.}

\item{krige_type, }{either 'ord' (ordinary; default), or 'univ' (universal)
kriging. Data for the respective kriging must be generated previously with
cell_krige().}

\item{plot_who, }{a vector of subject indexes as ordered within the STList, to
plot cells from. If NULL, will plot for all subjects.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{saveplot, }{logical indicating whether or not save plots in a PDF file.
The PDFs are saved in the working directory. Default is FALSE, meaning plots
are printed to console.}

\item{pvalues, }{logical indicating whether or not dots where a given cell was
predicted to be significantly abundant (p<0.05). Default is FALSE.}
}
\description{
Produces a kriging plot from cell scores.
}
\details{
This function produces a kriging plot for a series of cell names and spatial
arrays.
}
