% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_quilt.R
\name{plot_gene_quilt}
\alias{plot_gene_quilt}
\title{plot_gene_quilt: Gene expression at each spot from an spatial array}
\usage{
plot_gene_quilt(
  x = NULL,
  genes = NULL,
  plot_who = NULL,
  color_pal = "YlOrBr",
  data_type = "tr",
  purity = F,
  image = F,
  saveplot = NULL,
  inter = F,
  visium = T,
  ptsize = NULL
)
}
\arguments{
\item{x, }{an STList}

\item{genes, }{a vector of one or more gene names to plot.}

\item{plot_who, }{a vector of numbers indicating the spatial arrays to plot
genes from. Numbers follow the order of \code{names(x@counts)}. If NULL, will plot
all spatial arrays.}

\item{color_pal, }{a color scheme from 'khroma' or RColorBrewer.}

\item{data_type, }{one of 'tr' or 'raw', to plot transformed or raw counts
respectively.}

\item{purity, }{logical, whether tumor/stroma classes should be plotted.}

\item{saveplot, }{a file name specifying the name of a PDF file to write plots to.}

\item{inter, }{whether or not quilt plot should be converted to a Plotly plot.}

\item{visium, }{logical, whether or not the samples are from a Visium experiment.}
}
\value{
a list with plots.
}
\description{
Plot raw or transformed gene expression levels at each spot (quilt plot)
within a spatially-resolved transcriptomic array.
}
\details{
This function produces a quilt plot for a series of gene names and spatial
arrays within an STList. The function can also plot tumor/stroma classifications
from ESTIMATE deconvolution results.
}
\examples{
# In this example, melanoma is an STList.
# qplots <- plot_gene_quilt(melanoma, genes='CD74', plot_who=2, purity=T, visium=F)

}
