% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image.R
\name{plot_image}
\alias{plot_image}
\title{plot_image}
\usage{
plot_image(x = NULL, samples = NULL)
}
\arguments{
\item{x}{an STlist}

\item{samples}{a vector of numbers indicating the ST samples to plot, or their
sample names. If vector of numbers, it follow the order of \code{names(x@counts)}.
If NULL, the function plots all samples}
}
\value{
a list of plots
}
\description{
plot_image
}
\details{

}
