% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenovar.R
\name{plot_phenovar}
\alias{plot_phenovar}
\title{phenovar_plot: Plots a phenotype variable against gene expression spatial autocorrelation}
\usage{
plot_phenovar(
  x = NULL,
  phenovar = NULL,
  gene = NULL,
  cell = NULL,
  color_pal = "muted"
)
}
\arguments{
\item{x, }{an STList.}

\item{phenovar, }{a string indicating the name of the variable in the clinical
data frame. If NULL, uses sample}

\item{gene, }{the name of the gene to plot.}

\item{cell, }{the name of the cell type to plot.}

\item{color_pal, }{a string of a color palette from khroma.}
}
\value{
plist, a list with three plots.
}
\description{
Plots the value of a phenotype/clinical variable for each spatial
array against the spatial autocorrelation statistics of a given gene.
}
\details{
This function takes the name of a gene and calculates its spatial autocorrelation
statistics. Then generates a plot showing the variation of a phenotype/clinical
variable with respect to the Moran's I, Geary's C, and Getis-Ord Gi values of a
gene.
}
