% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_purity_quilt.R
\name{plot_purity_quilt}
\alias{plot_purity_quilt}
\title{plot_cell_quilt}
\usage{
plot_purity_quilt(
  x = NULL,
  cells = NULL,
  plot_who = NULL,
  color_pal = "YlOrBr",
  saveplot = F
)
}
\arguments{
\item{x, }{an STList with deconvolution matrices containing the cell names.}

\item{cells, }{a vector of cell names within a deconvolution matrix (one or
several) to plot.}

\item{plot_who, }{a vector of subject indexes as ordered within the STList, to
plot cells from. If NULL, will plot for all subjects.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{saveplot, }{logical indicating whether or not save plots in a PDF file.
The PDFs are saved in the working directory. Default is FALSE, meaning plots
are printed to console.}
}
\description{
Produces a quilt plot from cell scores.
}
\details{
This function produces a quilt plot for a series cell names and spatial
arrays.
}
