% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_quilt.R
\name{plot_QC_quilt}
\alias{plot_QC_quilt}
\title{plot_QC_quilt: Plot UMI counts at each spot from an spatial array}
\usage{
plot_QC_quilt(
  x = NULL,
  nreads = F,
  ngenes = F,
  pct_Expr = F,
  plot_who = NULL,
  spot_pctExpr = "^MT-",
  color_pal = "YlOrBr",
  purity = F,
  saveplot = NULL,
  inter = F,
  visium = T,
  ptsize = 0.5
)
}
\arguments{
\item{x}{an STList}

\item{nreads}{logical, if TRUE, plots the total counts per spot}

\item{ngenes}{logical, if TRUE, plots the number of UMI/genes per spot}

\item{pct_Expr}{logical, if TRUE, plots the percentage of counts from genes with
names matching a expression given by \code{spot_pctExpr}}

\item{plot_who, }{a vector of numbers indicating the spatial arrays to plot.Numbers
follow the order of \code{names(x@counts)}. If NULL, will plot all spatial arrays.}

\item{spot_pctExpr}{a expression to use with \code{pct_Expr}. By default '^MT-'.}

\item{color_pal, }{a color scheme from 'khroma' or RColorBrewer.}

\item{purity}{logical, whether tumor/stroma classes should be plotted.}

\item{saveplot, }{a file name specifying the name of a PDF file to write plots to.}

\item{inter, }{whether or not quilt plot should be converted to a Plotly plot.}

\item{visium, }{logical, whether or not the samples are from a Visium experiment.}

\item{ptsize, }{a number specifying the size of the points. Passed to \code{size} aesthetic.}
}
\value{
a list with plots.
}
\description{
Plots total counts, gene counts, or counts percentages at each spot
within a spatially-resolved transcriptomic array.
}
\details{
This function produces a quilt plot for total counts, gene/UMI counts, or percentage
of counts of a gene name matching a expression. This function is usfeul to explore
quality of the spots.
}
