% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_stats.R
\name{plot_QC_stats}
\alias{plot_QC_stats}
\title{plot_QCstats: Plots distribution of counts, gene number, or percents per sample}
\usage{
plot_QC_stats(
  x = NULL,
  qc = "percent_expr",
  plot_who = NULL,
  spot_pctExpr = "^MT-",
  ptsize = 0.5,
  ptalpha = 0.5,
  ptjitter = 0.2,
  saveplot = NULL,
  ylim = NULL
)
}
\arguments{
\item{x}{an STList}

\item{qc}{one of total_reads, total_genes, or percent_expr. The stat to be plotted}

\item{plot_who}{index from `x@counts``of the samples to be plotted}

\item{spot_pctExpr}{a expression to match gene names to calculate a percentage. By default '^MT-', to match mtDNA genes}

\item{ptsize}{the size of points in the plots}

\item{ptalpha}{the transparency of points in plots}

\item{saveplot}{File path to save a pdf with plots}
}
\description{
Generates plots with common quality control stats
}
\details{
This function returns a list of plots with one of three QC stats:
\itemize{
\item Total number of reads per spot
\item Total number of genes per spot
\item Percent of counts of genes matching \code{spot_pctExpr} per spot
}
}
