% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_samples.R
\name{pseudobulk_samples}
\alias{pseudobulk_samples}
\title{pseudobulk_samples: Aggregates counts into "pseudo bulk" samples}
\usage{
pseudobulk_samples(x = NULL, max_var_genes = 5000, calc_umap = F)
}
\arguments{
\item{x}{an STlist.}

\item{max_var_genes}{number of most variable genes (standard deviation) to use in
pseudobulk analysis}
}
\value{
an STlist with appended pseudobulk counts and PCA coordinates
}
\description{
Aggregates spot/cell counts into "pseudo bulk" samples for data exploration
}
\details{
This function takes an STlist and aggregates the spot/cell counts into "pseudo bulk"
counts by summing all counts from all cell/spots for each gene. Then performs
Principal Component Analysis (PCA) to explore non-spatial sample-to-sample
variation
}
\examples{
# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- system.file("extdata", 'melanoma_thrane', package="spatialGE")
count_files <- list.files(data_files, full.names=TRUE, pattern='counts')
coord_files <- list.files(data_files, full.names=TRUE, pattern='mapping')
clin_file <- list.files(data_files, full.names=TRUE, pattern='clinical')
melanoma <- STlist(rnacounts=count_files[c(1:4)], spotcoords=coord_files[c(1:4)], samples=clin_file) # Only first four samples
melanoma <- pseudobulk_samples(melanoma)
pseudobulk_pca_plot(melanoma)

}
