% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{quilt_p_purity}
\alias{quilt_p_purity}
\title{quilt_p_purity}
\usage{
quilt_p_purity(
  data_f = NULL,
  color_pal = "YlOrBr",
  leg_name = "",
  title_name = "",
  minvalue = minvalue,
  maxvalue = maxvalue,
  visium = T,
  ptsize = 0.5
)
}
\arguments{
\item{data_f, }{a data with four columns: x coordinates, y coordinates, values to be
plotted, and tumor/stroma classes.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{leg_name, }{a short name for the legend title.}

\item{title_name, }{a short name for the plot title.}

\item{minvalue, }{the minimum value of gene expression or cell score. Used for
standardization.}

\item{maxvalue, }{the maximum value of gene expression or cell score. Used for
standardization.}

\item{visium, }{whether or not to reverse axes for Visium slides.}

\item{ptsize, }{a number specifying the size of the points. Passed to \code{size} aesthetic.}
}
\value{
, a ggplot object.
}
\description{
Creates a quilt plot from ST data.
}
\details{
Function to produce a "quilt plot" from a data frame with four columns:
x coordinates, y coordinates, values (expression or cell scores), and tumor/stroma
classes. It also takes a color palette name from the 'khroma' or RColorBrewer
packages. Finally, it takes a name for the color legend title.
}
