% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{quilt_p_purity_bw}
\alias{quilt_p_purity_bw}
\title{quilt_p_purity_bw}
\usage{
quilt_p_purity_bw(data_f = NULL, title_name = "", visium = T, ptsize = 0.5)
}
\arguments{
\item{data_f, }{a data with four columns: x coordinates, y coordinates,values to be
plotted, and tumor/stroma classes. The third column is not used.}

\item{title_name, }{a short name for the plot title.}

\item{visium, }{whether or not to reverse axes for Visium slides.}

\item{ptsize, }{a number specifying the size of the points. Passed to \code{size} aesthetic.}
}
\value{
, a ggplot object.
}
\description{
Creates a tumor/stroma quilt plot from ST data without showing expression values.
}
\details{
Function to produce a "quilt plot" from a data frame with three columns:
x coordinates, y coordinates, and values (expression or cell scores). The data
frame has column names 'x_pos', 'y_pos', and 'values'.
}
