% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_helpers.R
\name{raise_err}
\alias{raise_err}
\title{raise_err: Raise error from exception database}
\usage{
raise_err(err_code = NULL, samplename = NULL)
}
\arguments{
\item{err_code}{error code to look up}

\item{samplename}{optional sample name to include in error message}
}
\description{
Stops with an error message from the spatialGE exception database
}
\keyword{internal}
