% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_helpers.R
\name{spatial_de}
\alias{spatial_de}
\title{spatial_de: Run spatial linear models with Matern covariance}
\usage{
spatial_de(expr_dat = NULL, non_sp_mods = NULL, annot_dict = NULL, verb = NULL)
}
\arguments{
\item{expr_dat}{a data frame with expression data and coordinates}

\item{non_sp_mods}{a list of non-spatial models to update with covariance structures}

\item{annot_dict}{a data frame mapping coded annotations to original annotation names}

\item{verb}{verbosity level (0, 1, or 2)}
}
\value{
a list of spatial models with convergence status
}
\description{
Fits spatial mixed models with Matern covariance structure to test
for spatial autocorrelation in gene expression between clusters
}
\keyword{internal}
