% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatial_metadata}
\alias{spatial_metadata}
\title{spatial_metadata: Prints the names of the available spot/cell annotations}
\usage{
spatial_metadata(x)
}
\arguments{
\item{x}{an STList object}
}
\description{
returns a character vector with the names of the annotations in the
\code{x@spatial_meta} slot.
}
