% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDeconv.R
\name{spatial_purity}
\alias{spatial_purity}
\title{spatial_purity}
\usage{
spatial_purity(x = NULL)
}
\arguments{
\item{x, }{a STList with normalized count matrices.}
}
\value{
x, an updated STList with ESTIMATE purity scores.
}
\description{
Applies ESTIMATE to ST data.
}
\details{
This function applies ESTIMATED to the stored transformed matrices in order to
obtain purity scores for each of the library/spots. The function runs on parallel
if unix system available.
}
