% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stat_plot.R
\name{spatial_stat_plot}
\alias{spatial_stat_plot}
\title{spatial_stat_plot: Plots association of samples variable and spatial statistics}
\usage{
spatial_stat_plot(
  x = NULL,
  samplevar = NULL,
  gene = NULL,
  cell = NULL,
  color_pal = "muted"
)
}
\arguments{
\item{x}{an STList.}

\item{samplevar}{a string indicating the name of the variable in the clinical
data frame. If NULL, uses sample}

\item{gene}{the name of the gene to plot.}

\item{cell}{the name of the cell type to plot.}

\item{color_pal}{a string of a color palette from khroma or RColorBrewer, or a
vector with colors with enough elements to plot categories.}
}
\value{
an STList containing the plots.
}
\description{
Plots the value of a sample/clinical variable against the spatial
autocorrelation statistics.
}
\details{
This function takes the name of a gene or deconvoluted cell score and calculates
its spatial autocorrelation statistics. Then generates a plot showing the variation
of a sample-associated or clinical variable with respect to the Moran's I, Geary's C,
and Getis-Ord Gi values of a gene.
}
\examples{
# In this example, melanoma is an STList.
# stat_p <- spatial_stat_plot(melanoma, samplevar='survival_months', gene='CD74')

}
