% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stat_plot.R
\name{spatial_stat_plot_gene}
\alias{spatial_stat_plot_gene}
\title{spatial_stat_plot_gene}
\usage{
spatial_stat_plot_gene(
  x = x,
  clinvar_vals = NULL,
  var = NULL,
  feature = NULL,
  color_pal = NULL
)
}
\arguments{
\item{x}{an STList}

\item{clinvar_vals}{a data frame with data frame to extract data for plot}

\item{var}{variable to plot from clinical data}

\item{feature}{the gene to plot statistics for}

\item{color_pal}{the name of a color palette}
}
\description{
spatial_stat_plot_gene
}
