% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_xcell.R
\name{spatial_xcell}
\alias{spatial_xcell}
\title{spatial_xcell}
\usage{
spatial_xcell(x = NULL)
}
\arguments{
\item{x, }{a STList with normalized count matrices.}
}
\value{
x, an updated STList with xCell scores.
}
\description{
Applies xCell to ST data.
}
\details{
This function applies xCell to the stored normalized matrices in order to obtain
cell scores for each of the library/spots. The results are stored as untransformed,
and square-root transformed scores. Stroma/Tumor scores are stored in a different
slot.
}
