% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDeconv.R
\name{spatialDeconv}
\alias{spatialDeconv}
\title{spatialDeconv: Deconvolution of ST data}
\usage{
spatialDeconv(x = NULL, method = "xcell")
}
\arguments{
\item{x}{an STList with normalized count matrices.}

\item{method}{the name of the method to deconvolute data. As of now, only xCell is supported.
If 'none', only ESTIMATE is applied.}
}
\value{
x an updated STList with deconvolution scores.
}
\description{
Applies gene expression deconvolution methods to data from spatial
arrays. Produces tumor/stroma classes via model-based clustering of ESTIMATE purity scores.
}
\details{
This function applies deconvolution methods (xCell and ESTIMATE, for now) to the stored
transformed count matrices in order to obtain cell scores for each of the library/spots.
The results are stored within the STList. The method ESTIMATE is automatically
applied regardless of the selected method.
}
\examples{
# In this example, melanoma is an STList.
# melanoma <- spatial_deconv(melanoma, method='xcell')

}
