% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTransform.R
\name{spatialTransform}
\alias{spatialTransform}
\title{spatialTransform: Data transformation of spatial transcriptomics data}
\usage{
spatialTransform(x = NULL, method = "log", scale_f = 10000)
}
\arguments{
\item{x}{an STList with raw count matrices.}

\item{method}{one of 'log' (default) or 'voom'.}

\item{scale_f, }{the scale factor used in logarithmic transformation.}
}
\value{
x an updated STList with transformed counts.
}
\description{
Applies data transformation methods to spatia transcriptomics experiments of an STList
}
\details{
This function takes an STList with raw counts and applies either logarithmic or
voom transformation after library size normalization. In the case of voom, library
size normalization applied is the trimmed mean of median values (TMM). If log,
counts are normalized using counts per spot. Logarithmic transformation also uses
a scaling factor (10^4) by default. The log, log(x+1) produces similar results to
Seurat's LogNormalize. The function works on parallel using "forking" (not in Windows OS).
}
\examples{
# In this example, melanoma is an STList.
# melanoma <- spatialTransform(melanoma)
# melanoma <- spatialTransform(melanoma, method='voom')

}
