% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STbulk_pca.R
\name{STbulk_pca}
\alias{STbulk_pca}
\title{STBulk_pca}
\usage{
STbulk_pca(x = NULL, clinvar = NULL)
}
\arguments{
\item{x, }{an STList.}

\item{clinvar, }{a string indicating the name of the variable in the clinical
data frame.}
}
\description{
Perform and plot a PCA of simulated bulk RNA-Seq ST data.
}
\details{
This function takes an STList and optionally the name of a clinical variable
and performs PCA on simulated bulk RNA-Seq from ST data. The counts are simulated
by summing all counts from an array for a given gene.
}
