% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust_core.R
\name{STclust_calculate_distances}
\alias{STclust_calculate_distances}
\title{STclust_calculate_distances: Calculate scaled distance matrices}
\usage{
STclust_calculate_distances(trcounts_df, coord_dat, dist_metric)
}
\arguments{
\item{trcounts_df}{a filtered expression matrix for one sample}

\item{coord_dat}{a data frame with spot coordinates for one sample}

\item{dist_metric}{distance metric to use}
}
\value{
a list with scaled distance matrices (expression + spatial)
}
\description{
Calculate and scale gene expression and spatial distance matrices for a single sample
}
\keyword{internal}
