% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust_core.R
\name{STclust_select_genes}
\alias{STclust_select_genes}
\title{STclust_select_genes: Select variable genes for clustering}
\usage{
STclust_select_genes(x, samples, topgenes, cores)
}
\arguments{
\item{x}{an STlist object with normalized expression data}

\item{samples}{vector of sample names to process}

\item{topgenes}{number of top variable genes to select per sample}

\item{cores}{number of cores for parallelization}
}
\value{
an STlist with updated \verb{@gene_meta} (VST variance) and \verb{@tr_counts} filtered to top genes
}
\description{
Selects top variable genes based on Seurat's VST method and filters expression data
}
\keyword{internal}
