% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust_core.R
\name{STclust_weight_distances}
\alias{STclust_weight_distances}
\title{STclust_weight_distances: Create weighted distance matrices}
\usage{
STclust_weight_distances(scaled_dists, ws)
}
\arguments{
\item{scaled_dists}{list of scaled distance matrices (expression + spatial)}

\item{ws}{vector of spatial weights to apply}
}
\value{
list of weighted distance matrices
}
\description{
Combine scaled expression and spatial distance matrices using user-specified weights
}
\keyword{internal}
