% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_nonspatial.R
\name{STdiff_fit_mm}
\alias{STdiff_fit_mm}
\title{STdiff_fit_mm: Fit non-spatial mixed models}
\usage{
STdiff_fit_mm(expr_data = NULL, combo = NULL, pairwise = NULL)
}
\arguments{
\item{expr_data}{a data frame (rows=spots/cells) with normalized expression data}

\item{combo}{a data frame with columns (samplename, meta1, meta2, gene) containing
all combinations of gene x cluster to test}

\item{pairwise}{whether to perform pairwise tests (TRUE) or reference-based (FALSE)}
}
\value{
a list containing non-spatial model results with p-values, log-fold changes,
and metadata
}
\description{
Fits non-spatial linear mixed models using spaMM for testing
differential expression between groups of spots/cells
}
\keyword{mixed}
\keyword{models}
\keyword{non-spatial}
