% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_core.R
\name{STdiff_fit_spatial}
\alias{STdiff_fit_spatial}
\title{STdiff_fit_spatial: Fit spatial mixed models for selected genes}
\usage{
STdiff_fit_spatial(
  x = NULL,
  prep = NULL,
  sp_topgenes = 0.2,
  cores = NULL,
  verbose = 1L
)
}
\arguments{
\item{x}{an STlist object}

\item{prep}{list from STdiff_select_genes containing combo_df, meta_dict, non_spatial_results}

\item{sp_topgenes}{proportion of top DE genes to fit spatial models on}

\item{cores}{number of cores for parallelization}

\item{verbose}{verbosity level}
}
\value{
list containing spatial model results
}
\description{
Fits spatial mixed models with Matern covariance structure for genes
selected from non-spatial testing
}
\keyword{internal}
