% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_nlme.R
\name{STdiff_old}
\alias{STdiff_old}
\title{STdiff: Differential gene expression analysis for spatial transcriptomics data}
\usage{
STdiff_old(
  x = NULL,
  samples = NULL,
  annot = NULL,
  ws = NULL,
  ks = "dtc",
  deepSplit = NULL,
  topgenes = 2000,
  pval_thr = 0.05,
  pval_adj = NULL,
  sp_topgenes = 0.2,
  verbose = F,
  cores = NULL
)
}
\arguments{
\item{x}{an STlist}

\item{samples}{an integer indicating the spatial samples to be included in the DE tests.
Numbers follow the order in \code{names(x@counts)}. Sample names are also allowed.
If NULL, performs tests in all samples}

\item{annot}{a column name in \code{x@spatial_meta} containing the groups/clusters to
be tested}

\item{ws}{the spatial weight used in STclust}

\item{ks}{the k value used in STclust, or \code{dtc} (default) for dynamicTreeCut clusters}

\item{deepSplit}{the deepSplit value if used in STclust}

\item{topgenes}{an integer indicating the top variable genes to select from each sample
based on variance (default=2000)}

\item{pval_thr}{Cut off of adjusted p-values to define differentially expressed genes from
non-spatial linear models. Default=0.05}

\item{pval_adj}{Method to adjust p-values from non-spatial models. Defaults to
\code{bonferroni}. Other options as available from \code{p.adjust}}

\item{sp_topgenes}{Proportion of differentially expressed genes from non-spatial
linear models (and controlled by \code{pval_thr}) to use in differential gene expression
analysis with spatial linear models. Default=0.2}

\item{cores}{Number of cores to use in parallelization. If \code{NULL}, the number of
cores is to use is detected automatically}
}
\value{
a data frame with results of differential gene expression analysis
}
\description{
Tests for differentially expressed genes using linear models with
exponential or spherical spatial covariance structures.
}
\details{
The method tests for differentially expressed genes between groups of spots/cells
(e.g., clusters) in a spatial transcriptomics sample (e.g., spots in a Visium array
or cells from a SMI slice). Specifically, the function tests for genes with
significantly higher or lower gene expression in one group of spots/cells with
respect to the rest of spots/cells in the sample. The method first runs linear
models on the genes to detect differentially expressed genes. Then on a subset of
differentially expressed genes, linear models (\code{sp_topgenes}) with spherical or
exponential covariance structures are fitted using the x,y coordinates of each
spot/cell. The function outputs the Akaike (AIC) and Bayesian Information Criterion (BIC)
for the spherical and exponential models to aid in selection of differentially expressed
genes under the spatial model with the best fit.
}
