% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_nonspatial.R
\name{STdiff_run_nonspatial}
\alias{STdiff_run_nonspatial}
\title{STdiff_run_nonspatial: Run non-spatial differential expression tests}
\usage{
STdiff_run_nonspatial(
  x = NULL,
  samples = NULL,
  annot = NULL,
  w = NULL,
  k = NULL,
  deepSplit = NULL,
  topgenes = 5000,
  pval_thr = 0.05,
  pval_adj = "fdr",
  test_type = "mm",
  clusters = NULL,
  pairwise = FALSE,
  verbose = 1L,
  cores = NULL
)
}
\arguments{
\item{x}{an STlist object}

\item{samples}{vector of sample names to test}

\item{annot}{column name in spatial_meta for cluster annotations}

\item{w}{spatial weight parameter (used if annot is NULL)}

\item{k}{number of clusters (used if annot is NULL)}

\item{deepSplit}{deepSplit parameter for dynamicTreeCut clusters}

\item{topgenes}{number of top variable genes to select}

\item{pval_thr}{p-value threshold for selecting DE genes}

\item{pval_adj}{p-value adjustment method}

\item{test_type}{type of test: 'mm', 't_test', or 'wilcoxon'}

\item{clusters}{optional vector of specific clusters to test}

\item{pairwise}{whether to perform pairwise tests (TRUE) or reference-based (FALSE)}

\item{verbose}{verbosity level (0, 1, or 2)}

\item{cores}{number of cores for parallelization}
}
\value{
list containing: combo_df, meta_dict, non_spatial_results, pval_thr, test_type, pairwise
}
\description{
Main entry point for non-spatial DE testing. Runs linear models,
t-tests, or Wilcoxon tests for selected genes between groups of spots/cells
}
\keyword{differential}
\keyword{expression}
\keyword{non-spatial}
