% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_core.R
\name{STdiff_select_genes}
\alias{STdiff_select_genes}
\title{STdiff_select_genes: Gene filtering and non-spatial differential testing}
\usage{
STdiff_select_genes(
  x = NULL,
  samples = NULL,
  annot = NULL,
  w = NULL,
  k = NULL,
  deepSplit = NULL,
  topgenes = 5000,
  pval_thr = 0.05,
  pval_adj = "fdr",
  test_type = "mm",
  clusters = NULL,
  pairwise = FALSE,
  verbose = 1L,
  cores = NULL
)
}
\arguments{
\item{x}{an STlist object}

\item{samples}{vector of sample names to test}

\item{annot}{column name in spatial_meta for cluster annotations}

\item{w}{spatial weight parameter (used if annot is NULL)}

\item{k}{number of clusters (used if annot is NULL)}

\item{deepSplit}{deepSplit parameter for dynamicTreeCut clusters}

\item{topgenes}{number of top variable genes to select}

\item{pval_thr}{p-value threshold for selecting DE genes}

\item{pval_adj}{p-value adjustment method}

\item{test_type}{type of test: 'mm', 't_test', or 'wilcoxon'}

\item{clusters}{optional vector of specific clusters to test}

\item{pairwise}{whether to perform pairwise tests}

\item{verbose}{verbosity level}

\item{cores}{number of cores for parallelization}
}
\value{
list containing: combo_df, meta_dict, non_spatial_results, pval_thr, test_type
}
\description{
Selects variable genes and performs non-spatial differential expression
testing (linear models, t-tests, or Wilcoxon tests)
}
\keyword{internal}
