% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_volcano.R
\name{STdiff_volcano}
\alias{STdiff_volcano}
\title{STdiff_volcano: Generates volcano plots from STdiff results}
\usage{
STdiff_volcano(
  x = NULL,
  samples = NULL,
  clusters = NULL,
  pval_thr = 0.05,
  color_pal = NULL
)
}
\arguments{
\item{x}{the output of \code{STdiff}}

\item{samples}{samples to create plots}

\item{clusters}{names of the clusters to generate comparisons}

\item{pval_thr}{the p-value threshold to color genes with differential expression}

\item{color_pal}{the palette to color genes by significance}
}
\value{
a list of ggplot objects
}
\description{
Generates volcano plots of differential expression results from STdiff
}
\details{
The function generated volcano plots (p-value vs. log-fold change) for
genes tested with \code{STdiff}. Colors can be customized to show significance from
spatial and non-spatial models
}
