% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich.R, R/STenrich_legacy.R
\name{STenrich}
\alias{STenrich}
\alias{STenrich_legacy}
\title{STenrich: Test for spatial enrichment of gene expression sets}
\usage{
STenrich(
  x,
  samples = NULL,
  gene_sets = NULL,
  score_type = "avg",
  reps = 1000,
  annot = NULL,
  domain = NULL,
  num_sds = 1,
  min_units = 20,
  min_genes = 5,
  pval_adj_method = "BH",
  seed = 12345,
  cores = NULL,
  verbose = TRUE
)

STenrich_legacy(
  x = NULL,
  samples = NULL,
  gene_sets = NULL,
  score_type = "avg",
  reps = 1000,
  annot = NULL,
  domain = NULL,
  num_sds = 1,
  min_units = 20,
  min_genes = 5,
  pval_adj_method = "BH",
  seed = 12345,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{an STlist with transformed gene expression}

\item{samples}{a vector with sample names or indexes to run analysis}

\item{gene_sets}{a named list of gene sets to test. The names of the list should
identify the gene sets to be tested}

\item{score_type}{Controls how gene set expression is calculated. The options are
the average expression among genes in a set ('avg'), or a GSEA score ('gsva'). The
default is 'avg'}

\item{reps}{the number of random samples to be extracted. Default is 1000 replicates}

\item{annot}{name of the annotation within \code{x@spatial_meta} containing the spot/cell
categories. Needs to be used in conjunction with \code{domain}}

\item{domain}{the domain to restrict the analysis. Must exist within the spot/cell
categories included in the selected annotation (i.e., \code{annot})}

\item{num_sds}{the number of standard deviations to set the minimum gene set
expression threshold. Default is one (1) standard deviation}

\item{min_units}{Minimum number of spots with high expression of a pathway for
that gene set to be considered in the analysis. Defaults to 20 spots or cells}

\item{min_genes}{the minimum number of genes of a gene set present in the data set
for that gene set to be included. Default is 5 genes}

\item{pval_adj_method}{the method for multiple comparison adjustment of p-values.
Options are the same as that of \code{p.adjust}. Default is 'BH'}

\item{seed}{the seed number for the selection of random samples. Default is 12345}

\item{cores}{the number of cores used during parallelization. If NULL (default),
the number of cores is defined automatically}

\item{verbose}{logical, whether to print text to console}
}
\value{
a list of data frames with the results of the test

list of data frames with p-values and adjusted p-values per sample
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Test for spatial enrichment of gene expression sets in spatial transcriptomics data

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Original implementation of STenrich for reproducibility with previous results

\strong{This is the legacy implementation from version 1.x.} Use \code{STenrich()} for new analyses.
}
\details{
The function performs a randomization test to assess if the sum of
distances between cells/spots with high expression of a gene set is lower than
the sum of distances among randomly selected cells/spots. The cells/spots are
considered as having high gene set expression if the average expression of genes in a
set is higher than the average expression plus \code{num_sds} times the standard deviation.
Control over the size of regions with high expression is provided by setting the
minimum number of cells/spots (\code{min_units}). This method is a modification of
the method devised by Hunter et al. 2021 (zebrafish melanoma study).
}
\examples{
\dontrun{
# Load TNBC test data
data_dir <- "tests/testthat/data/tnbc_bassiouni"
count_files <- list.files(data_dir, pattern='counts', full.names=TRUE)[1:2]
coord_files <- list.files(data_dir, pattern='mapping', full.names=TRUE)[1:2]
clin_file <- file.path(data_dir, "bassiouni_clinical.csv")

# Create STlist
tnbc <- STlist(rnacounts=count_files, spotcoords=coord_files, samples=clin_file)
tnbc <- transform_data(tnbc)

# Define gene sets (Hallmark pathways)
gene_sets <- list(
  "HALLMARK_APOPTOSIS" = c("CASP3", "CASP7", "BAX"),
  "HALLMARK_GLYCOLYSIS" = c("HK2", "PFKP", "LDHA")
)

# Run enrichment analysis
tnbc <- STenrich(tnbc, gene_sets=gene_sets, n_perm=100)
}

}
\seealso{
STenrich_legacy for reproducibility with original implementation
}
