% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich.R, R/STenrich_helpers.R
\name{STenrich_calculate_gs_gsva_score}
\alias{STenrich_calculate_gs_gsva_score}
\title{STenrich_calculate_gs_gsva_score: Calculate GSVA scores}
\usage{
STenrich_calculate_gs_gsva_score(
  x,
  pw_genes,
  gene_sets,
  min_genes,
  cores,
  verbose = TRUE
)

STenrich_calculate_gs_gsva_score(
  x,
  pw_genes,
  gene_sets,
  min_genes,
  cores,
  verbose = TRUE
)
}
\arguments{
\item{x}{an STlist with transformed gene expression (as DelayedArray)}

\item{pw_genes}{list of available genes per gene set}

\item{gene_sets}{a named list of gene sets to test}

\item{min_genes}{the minimum number of genes of a gene set present in data (default: 5)}

\item{cores}{the number of cores for parallelization}

\item{verbose}{verbosity level}
}
\value{
list of data frames with GSVA scores

list of data frames with GSVA scores
}
\description{
Calculate GSVA enrichment scores for gene sets

Calculate GSVA enrichment scores for gene sets
}
\keyword{internal}
