% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich_helpers.R
\name{STenrich_calculate_gs_mean_exp}
\alias{STenrich_calculate_gs_mean_exp}
\title{STenrich_calculate_gs_mean_exp: Calculate average gene set expression}
\usage{
STenrich_calculate_gs_mean_exp(x, combo, pw_genes, min_genes, cores)
}
\arguments{
\item{x}{an STlist with transformed gene expression (as DelayedArray)}

\item{combo}{data frame with combinations of samples and gene sets}

\item{pw_genes}{list of available genes per gene set}

\item{min_genes}{the minimum number of genes of a gene set present in data (default: 5)}

\item{cores}{the number of cores for parallelization}
}
\value{
list of data frames with average gene set expression
}
\description{
Calculate mean expression across genes in each gene set
}
\keyword{internal}
