% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich.R, R/STenrich_helpers.R
\name{STenrich_format_results}
\alias{STenrich_format_results}
\title{STenrich_format_results: Format final results}
\usage{
STenrich_format_results(pval_res, samples, pval_adj_method)

STenrich_format_results(pval_res, samples, pval_adj_method)
}
\arguments{
\item{pval_res}{data frame with p-values per sample and gene set}

\item{samples}{a vector with sample names to run analysis}

\item{pval_adj_method}{the method for multiple comparison adjustment (default: 'BH')}
}
\value{
list of data frames with p-values and adjusted p-values per sample

list of data frames with p-values and adjusted p-values per sample
}
\description{
Compile results and adjust p-values

Compile results and adjust p-values
}
\keyword{internal}
