% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich.R, R/STenrich_helpers.R
\name{STenrich_permutation_test}
\alias{STenrich_permutation_test}
\title{STenrich_permutation_test: Perform permutation testing}
\usage{
STenrich_permutation_test(
  result_df,
  coords_df,
  combo,
  pw_genes,
  samples,
  gene_sets,
  num_sds,
  min_units,
  reps,
  seed,
  cores,
  verbose
)

STenrich_permutation_test(
  result_df,
  coords_df,
  combo,
  pw_genes,
  samples,
  gene_sets,
  num_sds,
  min_units,
  reps,
  seed,
  cores,
  verbose
)
}
\arguments{
\item{result_df}{list of data frames with expression scores}

\item{coords_df}{list of coordinate matrices for each sample}

\item{combo}{data frame with combinations of samples and gene sets}

\item{pw_genes}{list of available genes per gene set}

\item{samples}{a vector with sample names to run analysis}

\item{gene_sets}{a named list of gene sets to test}

\item{num_sds}{number of standard deviations to set minimum gene set expression threshold (default: 1)}

\item{min_units}{Minimum number of spots with high expression (default: 20)}

\item{reps}{the number of random samples to be extracted (default: 1000)}

\item{seed}{the seed number for random sampling (default: 12345)}

\item{cores}{the number of cores for parallelization}

\item{verbose}{verbosity level}
}
\value{
list of data frames with p-values per sample

list of data frames with p-values per sample
}
\description{
Calculate observed distances and perform random permutations

Calculate observed distances and perform random permutations
}
\keyword{internal}
