% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich_helpers.R
\name{STenrich_prepare_data}
\alias{STenrich_prepare_data}
\title{STenrich_prepare_data: Prepare data for enrichment analysis}
\usage{
STenrich_prepare_data(x, samples, gene_sets, annot, domain, min_units)
}
\arguments{
\item{x}{an STlist with transformed gene expression}

\item{samples}{a vector with sample names to run analysis}

\item{gene_sets}{a named list of gene sets to test}

\item{annot}{name of the annotation within \code{x@spatial_meta} containing spot/cell categories}

\item{domain}{the domain to restrict the analysis}

\item{min_units}{Minimum number of spots with high expression (default: 20)}
}
\value{
list with prepared data including tissue_spots, coords_df, combo, pw_genes
}
\description{
Extract tissue spots, coordinates, and gene sets from STlist
}
\keyword{internal}
