% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich_helpers.R
\name{STenrich_validate_input}
\alias{STenrich_validate_input}
\title{STenrich_validate_input: Validate input parameters}
\usage{
STenrich_validate_input(
  x,
  samples,
  gene_sets,
  score_type,
  annot,
  domain,
  num_sds,
  min_units,
  min_genes,
  pval_adj_method
)
}
\arguments{
\item{x}{an STlist with transformed gene expression}

\item{samples}{a vector with sample names or indexes to run analysis}

\item{gene_sets}{a named list of gene sets to test}

\item{score_type}{Controls how gene set expression is calculated. Options: 'avg' or 'gsva'}

\item{annot}{name of the annotation within \code{x@spatial_meta} containing spot/cell categories}

\item{domain}{the domain to restrict the analysis}

\item{num_sds}{number of standard deviations to set minimum gene set expression threshold (default: 1)}

\item{min_units}{Minimum number of spots with high expression (default: 20)}

\item{min_genes}{the minimum number of genes of a gene set present in data (default: 5)}

\item{pval_adj_method}{the method for multiple comparison adjustment (default: 'BH')}
}
\value{
list with validated parameters
}
\description{
Check all input parameters and return validated values
}
\keyword{internal}
