% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_calculate_correlations}
\alias{STgradient_calculate_correlations}
\title{STgradient_calculate_correlations: Calculate correlations}
\usage{
STgradient_calculate_correlations(
  vargenes_expr,
  outs_dist2ref,
  robust,
  log_dist,
  sample_name
)
}
\arguments{
\item{vargenes_expr}{data.frame with expression and distance data}

\item{outs_dist2ref}{list of outlier barcodes per gene}

\item{robust}{logical whether to use robust regression}

\item{log_dist}{logical whether to log-transform distances}

\item{sample_name}{name of sample}
}
\value{
data.frame with correlation results
}
\description{
For each gene, fit linear models and calculate Spearman correlations
}
\keyword{internal}
