% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_cleanup}
\alias{STgradient_cleanup}
\title{STgradient_cleanup: Final cleanup}
\usage{
STgradient_cleanup(results_ls, zero_t, verbose)
}
\arguments{
\item{results_ls}{list of results}

\item{zero_t}{start time}

\item{verbose}{verbosity flag}
}
\value{
cleaned list of results
}
\description{
Remove empty samples and report timing
}
\keyword{internal}
