% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_core.R
\name{STgradient_core}
\alias{STgradient_core}
\title{STgradient_core: Core algorithm for spatial gradient analysis}
\usage{
STgradient_core(
  x,
  samples,
  annot,
  ref,
  exclude,
  out_rm,
  limit,
  distsumm,
  min_nb,
  robust,
  nb_dist_thr,
  log_dist,
  topgenes,
  cores,
  verbose
)
}
\arguments{
\item{x}{STlist object with transformed gene expression}

\item{samples}{Vector of sample names to process}

\item{annot}{Name of annotation column in @spatial_meta}

\item{ref}{Reference tissue domain}

\item{exclude}{Optional domain to exclude}

\item{out_rm}{Remove gene expression outliers (IQR method)}

\item{limit}{Limit analysis to spots within this distance threshold}

\item{distsumm}{Distance summary metric: "min" or "avg"}

\item{min_nb}{Minimum number of neighbors required}

\item{robust}{Use robust regression}

\item{nb_dist_thr}{Neighborhood distance threshold}

\item{log_dist}{Apply log transform to distances}

\item{topgenes}{Number of high-variance genes to test}

\item{cores}{Number of cores for parallelization}

\item{verbose}{Print progress messages}
}
\value{
List of data frames with correlation results
}
\description{
Core implementation of STgradient - calculates Spearman correlations between gene expression and distances to reference domain
}
\details{
Internal function - use STgradient() for public interface
}
