% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_detect_outliers}
\alias{STgradient_detect_outliers}
\title{STgradient_detect_outliers: Detect expression outliers}
\usage{
STgradient_detect_outliers(vargenes_expr, out_rm)
}
\arguments{
\item{vargenes_expr}{data.frame with expression and distance data}

\item{out_rm}{logical whether to detect outliers}
}
\value{
list with outlier barcodes per gene
}
\description{
Identify outlier spots using IQR-based method for each gene
}
\keyword{internal}
