% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_extract_expression}
\alias{STgradient_extract_expression}
\title{STgradient_extract_expression: Extract expression data}
\usage{
STgradient_extract_expression(x, sample_name, vargenes, dists_summ)
}
\arguments{
\item{x}{STlist object}

\item{sample_name}{sample name}

\item{vargenes}{vector of gene names}

\item{dists_summ}{distance summary data frame}
}
\value{
data frame with expression and distance data
}
\description{
Extract transformed expression for variable genes with distance data
}
\keyword{internal}
