% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_filter_neighbors}
\alias{STgradient_filter_neighbors}
\title{STgradient_filter_neighbors: Filter reference spots by neighbor count}
\usage{
STgradient_filter_neighbors(dist_tmp, ref_tmp, min_nb, nb_dist_thr)
}
\arguments{
\item{dist_tmp}{distance matrix}

\item{ref_tmp}{reference barcodes}

\item{min_nb}{minimum neighbors required}

\item{nb_dist_thr}{neighborhood distance threshold}
}
\value{
vector of reference barcodes to keep
}
\description{
Identify reference spots with sufficient neighbors
}
\keyword{internal}
