% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_format_results}
\alias{STgradient_format_results}
\title{STgradient_format_results: Format results}
\usage{
STgradient_format_results(dist_cor, distsumm)
}
\arguments{
\item{dist_cor}{correlation results data frame}

\item{distsumm}{distance summary metric}
}
\value{
formatted data frame
}
\description{
Format correlation results with proper column names
}
\keyword{internal}
