% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_identify_variable_genes}
\alias{STgradient_identify_variable_genes}
\title{STgradient_identify_variable_genes: Identify variable genes}
\usage{
STgradient_identify_variable_genes(x, sample_name, dists_summ, topgenes)
}
\arguments{
\item{x}{STlist object}

\item{sample_name}{sample name}

\item{dists_summ}{distance summary data frame}

\item{topgenes}{number of top genes to select}
}
\value{
vector of gene names
}
\description{
Select top variable genes for analysis
}
\keyword{internal}
