% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_prepare_distances}
\alias{STgradient_prepare_distances}
\title{STgradient_prepare_distances: Prepare distance data}
\usage{
STgradient_prepare_distances(x, sample_name, annot, ref, exclude)
}
\arguments{
\item{x}{STlist object}

\item{sample_name}{name of sample to process}

\item{annot}{annotation column name}

\item{ref}{reference cluster}

\item{exclude}{optional cluster to exclude}
}
\value{
list with distance data frames
}
\description{
Calculate and categorize distances for gradient analysis
}
\keyword{internal}
