% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_summarize_distances}
\alias{STgradient_summarize_distances}
\title{STgradient_summarize_distances: Summarize distances from reference}
\usage{
STgradient_summarize_distances(
  dist_tmp,
  nonref_tmp,
  ref_tmp,
  nbs_keep,
  distsumm,
  limit
)
}
\arguments{
\item{dist_tmp}{full distance matrix}

\item{nonref_tmp}{non-reference spot barcodes}

\item{ref_tmp}{reference spot barcodes}

\item{nbs_keep}{barcodes of reference spots with enough neighbors}

\item{distsumm}{distance summary method ('min' or 'avg')}

\item{limit}{optional distance limit}
}
\value{
data.frame with barcode and distance to reference
}
\description{
Calculate min or avg distance from non-reference spots to reference spots
}
\keyword{internal}
