% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_validate_input}
\alias{STgradient_validate_input}
\title{STgradient_validate_input: Validate input parameters}
\usage{
STgradient_validate_input(
  x,
  samples,
  topgenes,
  annot,
  ref,
  exclude,
  min_nb,
  nb_dist_thr,
  cores,
  verbose
)
}
\arguments{
\item{x}{an STlist object}

\item{samples}{vector of sample names to process}

\item{topgenes}{number of top variable genes}

\item{annot}{annotation column name}

\item{ref}{reference cluster}

\item{exclude}{optional exclude cluster}

\item{min_nb}{minimum number of neighbors}

\item{nb_dist_thr}{neighborhood distance threshold}

\item{verbose}{verbosity level}
}
\value{
list with validated inputs
}
\description{
Validate sample names, annotation columns, reference clusters, and other parameters
}
\keyword{internal}
