% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SThet_invdist_test.R
\name{SThet_invdist_test}
\alias{SThet_invdist_test}
\title{SThet_invdist_test: Computes spatial autocorrelation with statistical tests}
\usage{
SThet_invdist_test(
  x = NULL,
  genes = NULL,
  samples = NULL,
  method = "moran",
  k = NULL,
  overwrite = T,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{an STlist}

\item{genes}{a vector of gene names to compute statistics}

\item{samples}{the samples to compute statistics}

\item{method}{The spatial statistic(s) to estimate. It can be set to 'moran',
'geary' or both. Default is 'moran'}

\item{k}{the number of neighbors to estimate weights}

\item{overwrite}{logical indicating if previous statistics should be overwritten}

\item{cores}{the number of cores to use during computations}
}
\value{
an STlist containing spatial statistics
}
\description{
Alternative implementation using spdep::moran.test and spdep::geary.test
}
