% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SThet_legacy.R
\name{SThet_legacy}
\alias{SThet_legacy}
\title{SThet_legacy: Computes global spatial autocorrelation statistics on gene expression (legacy)}
\usage{
SThet_legacy(
  x = NULL,
  genes = NULL,
  samples = NULL,
  method = "moran",
  k = NULL,
  overwrite = TRUE,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{an STlist}

\item{genes}{a vector of gene names to compute statistics}

\item{samples}{the samples to compute statistics}

\item{method}{The spatial statistic(s) to estimate. It can be set to 'moran',
'geary' or both. Default is 'moran'}

\item{k}{the number of neighbors to estimate weights. By default NULL, meaning that
spatial weights will be estimated from Euclidean distances. If an positive integer is
entered, then the faster k nearest-neighbors approach is used. Please keep in mind
that estimates are not as accurate as when using the default distance-based method.}

\item{overwrite}{logical indicating if previous statistics should be overwritten.
Default to FALSE (do not overwrite)}

\item{cores}{integer indicating the number of cores to use during parallelization.
If NULL, the function uses half of the available cores at a maximum. The parallelization
uses \code{parallel::mclapply} and works only in Unix systems}

\item{verbose}{logical, whether to print text to console}
}
\value{
an STlist containing spatial statistics
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Computes the global spatial autocorrelation statistics Moran's I and/or
Geary's C for a set of genes
}
\details{
\strong{This is the legacy implementation from version 1.x.} Use \code{SThet} for new projects.

The function computes global spatial autocorrelation statistics (Moran's I and/or
Geary's C) for the requested genes and samples. Then computation uses the
package \code{spdep}. The calculated statistics are stored in the STlist, which can
be accessed with the \code{get_gene_meta} function. For visual comparative analysis,
the function \code{compare_SThet} can be used afterwards.
}
