% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STList.R
\name{STList_legacy}
\alias{STList_legacy}
\title{STlist (Legacy): Creation of STlist objects for spatial transcriptomics analysis}
\usage{
STList_legacy(
  rnacounts = NULL,
  spotcoords = NULL,
  samples = NULL,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{rnacounts}{the count data.}

\item{spotcoords}{the cell/spot coordinates.}

\item{samples}{the sample names/IDs.}

\item{cores}{integer indicating the number of cores to use.}

\item{verbose}{logical, whether to print text to console.}
}
\value{
an STlist object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Creates an STlist object from one or multiple spatial transcriptomic samples.
}
\details{
\strong{This function is superseded by the new \code{STList} implementation in version 2.0.0.}
It is preserved here for reproducibility of older workflows. Please use \code{STList} for new projects.
}
\examples{
\donttest{
# Using included melanoma example (Thrane et al.)
# Download example data set from spatialGE_Data
thrane_tmp = tempdir()
unlink(thrane_tmp, recursive=TRUE)
dir.create(thrane_tmp)
lk='https://github.com/FridleyLab/spatialGE_Data/raw/refs/heads/main/melanoma_thrane.zip?download='
tryCatch({ # In case data is not available from network
  download.file(lk, destfile=paste0(thrane_tmp, '/', 'melanoma_thrane.zip'), mode='wb')
  #' zip_tmp = list.files(thrane_tmp, pattern='melanoma_thrane.zip$', full.names=TRUE)
  unzip(zipfile=zip_tmp, exdir=thrane_tmp)
  # Generate the file paths to be passed to the STlist function
  count_files <- list.files(paste0(thrane_tmp, '/melanoma_thrane'),
                            full.names=TRUE, pattern='counts')
  coord_files <- list.files(paste0(thrane_tmp, '/melanoma_thrane'),
                            full.names=TRUE, pattern='mapping')
  clin_file <- list.files(paste0(thrane_tmp, '/melanoma_thrane'),
                          full.names=TRUE, pattern='clinical')
  # Create STlist
  library('spatialGE')
  melanoma <- STlist(rnacounts=count_files,
                     spotcoords=coord_files,
                     samples=clin_file)
  melanoma
}, error = function(e) {
  message("Could not run example. Are you connected to the internet?")
  return(NULL)
})
}

}
\seealso{
\code{\link{STList}}
}
