% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_STlist.R
\name{summarize_STlist}
\alias{summarize_STlist}
\title{summarize_STlist: Generates a data frame with summary statistics}
\usage{
summarize_STlist(x = NULL)
}
\arguments{
\item{x}{an STlist}
}
\value{
a data frame
}
\description{
Produces a data frame with counts per gene and counts per ROI/spot/cell
}
\details{
The function creates a table with counts per gene and counts per region of interest (ROI),
spot, or cell in the samples stored in the STlist
}
\examples{
# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- system.file("extdata", 'melanoma_thrane', package="spatialGE")
count_files <- list.files(data_files, full.names=TRUE, pattern='counts')
coord_files <- list.files(data_files, full.names=TRUE, pattern='mapping')
clin_file <- list.files(data_files, full.names=TRUE, pattern='clinical')
melanoma <- STlist(rnacounts=count_files[c(1,2)], spotcoords=coord_files[c(1,2)], samples=clin_file) # Only first two samples
summarize_STlist(melanoma)

}
