% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voom_norm.R
\name{voom_norm}
\alias{voom_norm}
\title{voom_norm: Normalization of ST arrays}
\usage{
voom_norm(x = NULL)
}
\arguments{
\item{x, }{an STList with raw count matrices.}
}
\value{
x, an STList with normalized counts.
}
\description{
Applies limma-voom normalization to spatial transcriptomics arrays data.
}
\details{
This function takes an STList and normalize the count data matrices within it
in two steps. In the first step, (edgeR) normalization factors are used to scale
the counts from each library/spot. In the second step, limma-voom normalization
is applied. The resulting normalized count matrix is stored in the voom_counts
slot of the STList. The function also calculates gene-wise standard deviation from
the normalized counts and stores them in the gene_stdev slot.

NOTE1: ADD MEAN-VAR PLOT TO THIS FUNCTION??
}
