% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xcell_names}
\alias{xcell_names}
\title{xcell_names: Prints xCell cell type names}
\usage{
xcell_names()
}
\value{
The xCell cell type names after cleaning.
}
\description{
Prints the available cell types as provided by xCell deconvolution.
}
\details{
Shows the names of cell types that can be used in other spatialGE functions.
This function access the xCell cell type names after cleaning via the janitor
package.
}
\examples{
# xcell_names()

}
