% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_purity.R
\name{cluster_purity}
\alias{cluster_purity}
\title{cluster_purity}
\usage{
cluster_purity(x = NULL)
}
\arguments{
\item{x, }{an STList with ESTIMATE scores.}
}
\value{
x, an STList with tumor/stroma classifications.
}
\description{
Perform model-based clustering to define stromal/tumor clusters from
ESTIMATE purity scores.
}
\details{
Takes ESTIMATE tumor purity scores and performs model-based clustering. The function
only evaluates k=2, to separate between likely tumor and stromal spots.
}
