% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_STspot.R
\name{cluster_STspot}
\alias{cluster_STspot}
\title{cluster_STspot}
\usage{
cluster_STspot(x = NULL, who = NULL)
}
\arguments{
\item{x, }{an STList.}

\item{who, }{an integer indicating the spatial array to be analyzed.}
}
\description{
Perform {INSERT CLUSTER ALGORITHM HERE :|} of spots within ST array.
}
\details{
yadda yadda yadda
}
