% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_cores}
\alias{count_cores}
\title{Get the number of cores to use in parallel as a function of the number of
data to process. It will not yield a higher number of cores than half of the
total available cores. Will default to 1 core if not Unix}
\usage{
count_cores(n)
}
\arguments{
\item{n}{an integer representing the number of units to process}
}
\value{
the number of cores to use
}
\description{
Get the number of cores to use in parallel as a function of the number of
data to process. It will not yield a higher number of cores than half of the
total available cores. Will default to 1 core if not Unix
}
