% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_krige.R
\name{deconv_krige}
\alias{deconv_krige}
\title{deconv_krige}
\usage{
deconv_krige(x = NULL, cells = "top", univ = F, res = 0.2, who = NULL)
}
\arguments{
\item{x, }{an STList with transformed xCell scores.}

\item{cells, }{a vector of cell names or 'top'. If 'top' (default), kriging for the 10
cells with highest standard deviation is estimated.}

\item{univ, }{a logical stating whether or not to perform universal or ordinary kriging.
Default is FALSE (ordinary kriging).}

\item{res, }{a double to adjust the resolution of the plot. Fractions of 1 lead to
more resolution. Default is res=0.2.}

\item{who, }{the spatial arrays for which krginig will be performed. If NULL (Default),
all arrays are kriged.}
}
\value{
x, a STList including spatial interpolations.
}
\description{
Performs spatial interpolation ('kriging') of deconvoluted cell
scores in spatially-resolved transcriptomics data.
}
\details{
This function takes a STList and a vector of xCell cell names, or the token 'top' for
the 10 cells with the highest standard deviation. It also calculates spatial heterogeneity
statistics for the cell scores. The function can perform ordinary or universal kriging.
The result can be plotted using the plot_cell_krige() function. The stroma score from
xCell is automatically kriged.
}
