% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_utils.R
\name{detect_input_source}
\alias{detect_input_source}
\title{Detect Input Source}
\usage{
detect_input_source(rnacounts = NULL, spotcoords = NULL, samples = NULL)
}
\description{
Analyzes arguments to determine the Input Source strategy
}
