% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_generic.R
\name{dispatch_ingest.source_generic}
\alias{dispatch_ingest.source_generic}
\title{Ingest Generic Data (DataFrames, Lists, Text Files)}
\usage{
\method{dispatch_ingest}{source_generic}(source)
}
\arguments{
\item{source}{InputSource object of type 'generic' or 'list'}
}
\description{
Ingest Generic Data (DataFrames, Lists, Text Files)
}
