% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_visium.R
\name{dispatch_ingest.source_h5_10x}
\alias{dispatch_ingest.source_h5_10x}
\title{Ingest Generic H5 Details (Visium or Xenium or 10x)}
\usage{
\method{dispatch_ingest}{source_h5_10x}(source)
}
\arguments{
\item{source}{InputSource object of type 'h5_10x'}
}
\description{
Ingest Generic H5 Details (Visium or Xenium or 10x)
}
