% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_generic.R
\name{dispatch_ingest.source_seurat}
\alias{dispatch_ingest.source_seurat}
\title{Ingest Seurat Object}
\usage{
\method{dispatch_ingest}{source_seurat}(source)
}
\arguments{
\item{source}{InputSource object of type 'seurat'}
}
\description{
Ingest Seurat Object
}
