% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_xenium.R
\name{dispatch_ingest.source_xenium}
\alias{dispatch_ingest.source_xenium}
\title{Ingest Xenium Data}
\usage{
\method{dispatch_ingest}{source_xenium}(source)
}
\arguments{
\item{source}{InputSource object of type 'xenium'}
}
\description{
Ingest Xenium Data
}
