% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{filter_data: Removes spots, genes, or samples}
\usage{
filter_data(
  x = NULL,
  spot_minreads = NULL,
  spot_maxreads = NULL,
  spot_mingenes = NULL,
  spot_maxgenes = NULL,
  spot_minpct = NULL,
  spot_maxpct = NULL,
  gene_minreads = NULL,
  gene_maxreads = NULL,
  gene_minspots = NULL,
  gene_maxspots = NULL,
  gene_minpct = NULL,
  gene_maxpct = NULL,
  who = NULL,
  rm_tissue = NULL,
  rm_genes = NULL,
  rm_partialgenes = NULL,
  spot_pctExpr = "^MT-"
)
}
\arguments{
\item{x}{an STList}

\item{spot_minreads}{the minimum number of total reads for a spot to be retained}

\item{spot_maxreads}{the maximum number of total reads for a spot to be retained}

\item{spot_mingenes}{the minimum number of non-zero counts for a spot to be retained}

\item{spot_maxgenes}{the maximum number of non-zero counts for a spot to be retained}

\item{spot_minpct}{the minimum percentage of counts for features defined by \code{spot_pctexpr}
for a spot to be retained.}

\item{spot_maxpct}{the maximum percentage of counts for features defined by \code{spot_pctexpr}
for a spot to be retained.}

\item{gene_minreads}{the minimum number of total reads for a gene to be retained}

\item{gene_maxreads}{the maximum number of total reads for a gene to be retained}

\item{gene_minspots}{he minimum number of spots with non-zero counts for a gene to be retained}

\item{gene_maxspots}{the maximum number of spots with non-zero counts for a gene to be retained}

\item{gene_minpct}{the minimum percentage of spots with non-zero counts for a gene to be retained}

\item{gene_maxpct}{the maximum percentage of spots with non-zero counts for a gene to be retained}

\item{who}{vector of integers (as in \code{x@counts}) or sample names to perform filtering}

\item{rm_tissue}{vector of integers (as in \code{x@counts}) or sample names to remove from STList}

\item{rm_genes}{vector of gene names to remove from STList}

\item{rm_partialgenes}{a expression to match gene names. Useful to remove entire gene classes (for example mtDNA '^MT-')}

\item{spot_pctExpr}{a expression to use with \code{spot_minpct}. By default '^MT-'.}
}
\description{
Provides methods to remove spots and genes or entire samples.
}
\details{
This function provides methods to filter an STList. It can remove spots or genes
based on absolute counts or percentages. Users can input a expression to be matched
with gene names to calculate percentages (for example \%mtDNA genes).
Note that the function removes counts raw counts and spatial information. Transformed
counts or other results within the STList remain unaffected.
}
