% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_spatialstats.R
\name{gene_geary_C}
\alias{gene_geary_C}
\title{gene_geary_C}
\usage{
gene_geary_C(x = NULL, genes = NULL, who = NULL)
}
\arguments{
\item{x, }{an STList with normalized gene counts.}

\item{genes, }{a vector with gene names in the normalized count matrix.}

\item{who, }{the index of the spatial array for which the statistic
will be calculated.}
}
\value{
x, a STList including the values corresponding to Geary's C for each
gene in genes.
}
\description{
Calculates Geary's C from ST data.
}
\details{
This function takes an STList and a vector with HUGO gene names and returns
Geary's C for each element of the vector.
}
