% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_spatialstats.R
\name{gene_getis_Gi}
\alias{gene_getis_Gi}
\title{gene_getis_Gi}
\usage{
gene_getis_Gi(x = NULL, genes = NULL, who = NULL)
}
\arguments{
\item{x, }{an STList with normalized gene counts.}

\item{genes, }{a vector with gene names in the normalized count matrix.}

\item{who, }{the indexes of the spatial arrays for which the statistic
will be calculated.}
}
\value{
x, a STList including the values corresponding to Getis-Ord Gi for each
gene in genes.
}
\description{
Calculates Getis-Ord Gi C from ST data.
}
\details{
This function takes an STList and a vector with HUGO gene names and returns
Getis-Ord Gi for each element of the vector.
}
