% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_krige.R
\name{gene_krige}
\alias{gene_krige}
\title{gene_krige: Spatial interpolation of gene expression}
\usage{
gene_krige(x = NULL, genes = "top", univ = F, res = NULL, who = NULL)
}
\arguments{
\item{x, }{an STList with normalized counts}

\item{genes, }{a vector of HUGO names or 'top'. If 'top' (default), kriging for the 10
genes with highest standard deviation is estimated.}

\item{univ, }{a logical stating whether or not to perform universal kriging.
The default is FALSE, meaning that ordinary kriging will be performed.}

\item{res, }{a double to adjust the resolution of the plot. Fractions of 1 lead to
more resolution. Default is res=0.2.}

\item{who, }{the spatial arrays for which krginig will be performed. If NULL (Default),
all arrays are kriged.}
}
\value{
x, a STList including spatial interpolations.
}
\description{
Performs spatial interpolation ('kriging') of normalized gene
counts in spatially-resolved transcriptomics data.
}
\details{
This function takes a STList and a vector of gene names, or the token 'top' for the 10
genes with the highest standard deviation. It also calculates spatial heterogeneity
statistics for the genes counts. The function can perform ordinary or universal kriging.
The result can be plotted using the plot_gene_krige() function.
}
