% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_visium.R
\name{ingest_visium_mex_dir}
\alias{ingest_visium_mex_dir}
\title{Internal MEX Directory Reader}
\usage{
ingest_visium_mex_dir(
  mex_dir,
  coords_path = NULL,
  sample_name = NULL,
  original_root = NULL
)
}
\description{
Internal MEX Directory Reader
}
