% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_xenium.R
\name{ingest_xenium_mex_dir}
\alias{ingest_xenium_mex_dir}
\title{Internal Xenium MEX Directory Reader}
\usage{
ingest_xenium_mex_dir(
  mex_dir,
  coords_path = NULL,
  sample_name = NULL,
  original_root = NULL
)
}
\description{
Internal Xenium MEX Directory Reader
}
