% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconv_krige.R
\name{plot_deconv_krige}
\alias{plot_deconv_krige}
\title{plot_deconv_krige: Visualize spatial interpolation of deconvolution scores}
\usage{
plot_deconv_krige(
  x = NULL,
  cells = NULL,
  plot_who = NULL,
  color_pal = "YlOrBr",
  method = "xCell",
  saveplot = F,
  pvalues = F,
  purity = F,
  visium = T,
  ptsize = 0.5
)
}
\arguments{
\item{x}{an STList with kriging objects for the cells selected.}

\item{cells, }{a vector of cell names (one or several) within a deconvolution
matrix to plot. If 'top', the 10 cell types with highest standard deviation
from each spatial array ar plotted.}

\item{plot_who}{a vector of subject indexes as ordered within the STList, to
plot cells from. Numbers follow the order in \code{names(x@counts)}. If NULL, will
plot for all subjects.}

\item{color_pal}{a color scheme from 'khroma' or RColorBrewer.}

\item{saveplot, }{a file name specifying the name of a PDF file to write plots to.}

\item{pvalues, }{logical indicating whether or not dots where a given cell was
predicted to be significantly abundant (p<0.05). Default is FALSE.}

\item{purity, }{logical, whether or not to annotate tumor spots based on
ESTIMATE tumor purity scores.}

\item{visium, }{whether or not to reverse axes for Visium slides.}

\item{image}{logical, whether to print the image stored for the spatial arrays}
}
\value{
a list with plots.
}
\description{
Produces a spatial interpolation surface from kriging of gene expression
deconvolution scores.
}
\details{
This function produces a spatial interpolation surface plot for a series of
cell names and spatial arrays.
}
\examples{
# In this example, melanoma is an STList.
# kplots <- plot_deconv_krige(melanoma,  cells=c('b_cells', 'i_dc'), plot_who=2, visium=F)

}
