% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconv_quilt.R
\name{plot_deconv_quilt}
\alias{plot_deconv_quilt}
\title{plot_deconv_quilt: Deconvolution scores at each spot from an spatial array}
\usage{
plot_deconv_quilt(
  x = NULL,
  cells = NULL,
  plot_who = NULL,
  color_pal = "YlOrBr",
  purity = F,
  saveplot = F,
  scaled = F,
  method = "xcell",
  visium = T
)
}
\arguments{
\item{x, }{an STList with deconvolution matrices containing the cell names.}

\item{cells, }{a vector of cell names within a deconvolution matrix (one or
several) to plot.}

\item{plot_who, }{a vector of subject indexes as ordered within the STList, to
plot cells from. If NULL, will plot for all subjects.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.'}

\item{purity, }{logical, whether tumor/stroma classes should be plotted.}

\item{saveplot, }{logical indicating whether or not save plots in a PDF file.
The PDFs are saved in the working directory. Default is FALSE, meaning plots
are printed to console.}

\item{scaled, }{logical, indicating if expression values should be scaled with
respect to the highest value among all genes to plot. WARNING: Color legends
are not scaled between plots, but values are.}

\item{method, }{the deconvolutionn method to used estimate the cell scores.}

\item{visium, }{whether or not to reverse axes for Visium slides.}
}
\value{
qplist, a list with plots.
}
\description{
Plot cell type deconvolution scores at each spot (quilt plot) within a
spatially-resolved transcriptomic array.
}
\details{
This function produces a quilt plot for a series cell names and spatial
arrays within an STList. The function can also plot tumor/stroma classifications
if ESTIMATE deconvolution results are present in the STList.
}
