% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_krige.R
\name{plot_gene_krige}
\alias{plot_gene_krige}
\title{plot_gene_krige}
\usage{
plot_gene_krige(
  x = NULL,
  genes = NULL,
  krige_type = "ord",
  plot_who = NULL,
  color_pal = "YlOrBr",
  saveplot = F
)
}
\arguments{
\item{x, }{an STList with kriging objects for the genes selected.}

\item{genes, }{a vector of gene names (one or several) to plot.}

\item{krige_type, }{either 'ord' (ordinary; default), or 'univ' (universal)
kriging. Data for the respective kriging must be generated previously with
cell_krige().}

\item{plot_who, }{a vector of subject indexes as ordered within the STList, to
plot genes from. If NULL, will plot for all subjects.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{saveplot, }{logical indicating whether or not save plots in a PDF file.
The PDFs are saved in the working directory. Default is FALSE, meaning plots
are printed to console.}
}
\description{
Produces a kriging plot from ST data.
}
\details{
This function produces a kriging plot for a series of HUGO gene names and
spatial arrays.
}
