% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_krige.R
\name{plot_gene_krige}
\alias{plot_gene_krige}
\title{plot_gene_krige: Visualize transcriptomic surfaces}
\usage{
plot_gene_krige(
  x = NULL,
  genes = NULL,
  plot_who = NULL,
  color_pal = "YlOrBr",
  purity = F,
  image = F,
  saveplot = NULL,
  visium = T,
  ptsize = 0.5
)
}
\arguments{
\item{x}{an STList with kriging objects for the genes selected.}

\item{genes}{a vector of gene names (one or several) to plot.  If 'top', the 10
genes with highest standard deviation from each spatial array ar plotted.}

\item{plot_who, }{a vector of numbers indicating the spatial arrays to plot
genes from. Numbers follow the order in \code{names(x@counts)}. If NULL, will plot
all spatial arrays.}

\item{color_pal}{a color scheme from 'khroma' or RColorBrewer.}

\item{purity}{logical, whether or not annotate tumor spots based on
ESTIMATE tumor purity scores.}

\item{image}{logical, whether to print the image stored for the spatial arrays}

\item{saveplot, }{a file name specifying the name of a PDF file to write plots to.}

\item{visium}{whether or not to reverse axes for Visium slides.}
}
\value{
a list with plots.
}
\description{
Produces a transcriptomic surface from kriging interpolation of ST data.
}
\details{
This function produces a transcriptomic surface plot for a series of HUGO gene names and
spatial arrays.
}
\examples{
# In this example, melanoma is an STList.
# kplots <- plot_gene_krige(melanoma, genes=c('CD74', 'SOX10'), plot_who=3, visium=F)

}
