% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_quilt.R
\name{plot_gene_quilt}
\alias{plot_gene_quilt}
\title{plot_gene_quilt}
\usage{
plot_gene_quilt(
  x = NULL,
  genes = NULL,
  plot_who = NULL,
  color_pal = "YlOrBr",
  saveplot = F
)
}
\arguments{
\item{x, }{an STList with voom-norm counts.}

\item{genes, }{a vector of gene names (one or several) to plot.}

\item{plot_who, }{a vector of subject indexes as ordered within the STList, to
plot genes from. If NULL, will plot for all subjects.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{saveplot, }{logical indicating whether or not save plots in a PDF file.
The PDFs are saved in the working directory. Default is FALSE, meaning plots
are printed to console.}
}
\description{
Produces a quilt plot from ST data.
}
\details{
This function produces a quilt plot for a series of HUGO gene names and spatial
arrays
}
