% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image.R
\name{plot_image}
\alias{plot_image}
\title{plot_image: Generate a ggplot object of the tissue image}
\usage{
plot_image(x = NULL, samples = NULL)
}
\arguments{
\item{x}{an STlist}

\item{samples}{a vector of numbers indicating the ST samples to plot, or their
sample names. If vector of numbers, it follow the order of \code{names(x@counts)}.
If NULL, the function plots all samples}
}
\value{
a list of plots
}
\description{
Creates ggplot objects of the tissue images when available within the STlist
}
\details{
If the STlist contains tissue images in the \verb{@misc} slot, the \code{plot_image} function
can be used to generate ggplot objects. These ggplot objects can be plotted next to
quilt plots (\code{STplot} function) for comparative analysis.
}
