% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_STclusters.R
\name{plot_STclusters}
\alias{plot_STclusters}
\title{plot_STclusters: Plot cluster memberships of ST spots}
\usage{
plot_STclusters(
  x,
  plot_who = NULL,
  ks = NULL,
  weights = NULL,
  purity = F,
  color_pal = "light",
  visium = T,
  ptsize = NULL
)
}
\arguments{
\item{x}{an STList with hierarchical cluster memberships.}

\item{plot_who}{an integer indicating the spatial array(s) to be plotted. Numbers
follow the order of \code{names(x@counts)}. If NULL, will plot all spatial arrays.}

\item{ks}{the k values to plot}

\item{weights}{the spatial weights to plot}

\item{purity}{logical, whether or not annotate tumor/stroma.}

\item{color_pal}{a string of a color palette from khroma or RColorBrewer, or a
vector with colors with enough elements to plot categories.}

\item{visium}{whether or not to reverse axes for Visium slides.}
}
\value{
a list with the requested plots.
}
\description{
Generates a plot of the location of spots within an spatial array,
and colors them according to spatially-weighted hierarchical clustering assignments.
}
\details{
The function takes an STList with cluster memberships and plots the spots with
colors indicating the cluster they belong to. Optionally, the user can annotate
tumor/stroma compartments if ESTIMATE scores are available.
}
\examples{
# In this example, melanoma is an STList.
# cluster_p <- plot_STclusters(melanoma, purity=T, plot_who=c(2,3), visium=F)

}
