% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_STclusters.R
\name{plot_STclusters}
\alias{plot_STclusters}
\title{plot_STclusters: Plot cluster memberships of ST spots}
\usage{
plot_STclusters(x, plot_who = NULL, purity = F, color_pal = "light")
}
\arguments{
\item{x, }{an STList with hierarchical cluster memberships.}

\item{purity, }{logical, whether or not annotate tumor/stroma.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'light'.}

\item{who, }{an integer indicating the spatial array(s) to be plotted.}
}
\description{
Generates a plot of the location of spots within an spatial array,
and colors them according to spatially-weighted hierarchical clustering assignments.
}
\details{
The function takes an STList with cluster memberships and plots the spots with
colors indicating the cluster they belong to. Optionally, the user can annotate
tumor/stroma compartments if ESTIMATE scores are available.
}
