% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_samples.R
\name{pseudobulk_samples}
\alias{pseudobulk_samples}
\title{pseudobulk_samples: Aggregates counts into "pseudo bulk" samples}
\usage{
pseudobulk_samples(x = NULL, max_var_genes = 5000)
}
\arguments{
\item{x}{an STlist.}

\item{max_var_genes}{number of most variable genes (standard deviation) to use in
pseudobulk analysis}
}
\value{
an STlist with appended pseudobulk counts and PCA coordinates
}
\description{
Aggregates spot/cell counts into "pseudo bulk" samples for data exploration
}
\details{
This function takes an STlist and aggregates the spot/cell counts into "pseudo bulk"
counts by summing all counts from all cell/spots for each gene. Then performs
Principal Component Analysis (PCA) to explore non-spatial sample-to-sample
variation
}
\examples{
# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- list.files(system.file("extdata", package="spatialGE"), recursive=T, full.names=T)
count_files <- grep("counts", data_files, value=T)
coord_files <- grep("mapping", data_files, value=T)
clin_file <- grep("thrane_clinical", data_files, value=T)
melanoma <- STlist(rnacounts=count_files[c(1:4)], spotcoords=coord_files[c(1:4)], samples=clin_file) # Only first two samples
melanoma <- pseudobulk_samples(melanoma)
pseudobulk_pca_plot(melanoma)

}
