% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt_p_purity.R
\name{quilt_p_purity}
\alias{quilt_p_purity}
\title{quilt_p_purity}
\usage{
quilt_p_purity(
  data_f = NULL,
  color_pal = "YlOrBr",
  leg_name = "",
  title_name = ""
)
}
\arguments{
\item{data_f, }{a data with three columns: x coordinates, y coordinates, and
the values to be plotted.}

\item{color_pal, }{a scheme from 'khroma'. Default is 'YlOrBr'.}

\item{leg_name, }{a short name for the legend title.}

\item{title_name, }{a short name for the plot title.}
}
\value{
, a ggplot object.
}
\description{
Creates a quilt plot from ST data.
}
\details{
Function to produce a "quilt plot" from a data frame with three columns:
x coordinates, y coordinates, and values (expression or cell scores). The data
frame has column names 'x_pos', 'y_pos', and 'values'. It also takes a color
palette name from the 'khroma' package. Finally, it takes a name for the color
legend title.
}
