% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\name{show,STlist-method}
\alias{show,STlist-method}
\title{show: Prints overview of STList oject.}
\usage{
\S4method{show}{STlist}(object)
}
\arguments{
\item{object}{an STList object to show summary from.}
}
\description{
Prints overview/summary of STList oject.
}
\details{
This function takes an STList and prints a the number of spatial arrays in that
object and other information about the object.
}
