% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconv.R
\name{spatial_deconv}
\alias{spatial_deconv}
\title{spatial_deconv}
\usage{
spatial_deconv(x = NULL, method = "xcell")
}
\arguments{
\item{x, }{a STList with normalized count matrices. If 'none', only ESTIMATE is
applied.}
}
\value{
x, an updated STList with deconvolution scores.
}
\description{
Applies deconvolution methods to ST data.
}
\details{
This function applies deconvolution methods (xCell by default) to the stored
normalized matrices in order to obtain cell scores for each of the library/spots.
The results are stored as untransformed, and square-root transformed scores.
The method ESTIMATE is automatically applied regardless of the selected method.
}
