% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatial_metadata}
\alias{spatial_metadata}
\title{spatial_metadata: Prints the names of the available spot/cell annotations}
\usage{
spatial_metadata(x)
}
\arguments{
\item{x}{an STList object}
}
\value{
a list of character vectors containing the column names of \code{x@spatial_meta}
}
\description{
returns the names of the annotations in the \code{x@spatial_meta} slot.
}
