% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust.R
\name{STclust}
\alias{STclust}
\title{STclust: Detect clusters on ST arrays}
\usage{
STclust(
  x = NULL,
  weights = 0.025,
  pcs = NULL,
  vperc = NULL,
  method = "ward.D",
  ks = "dtc",
  topgenes = 3000,
  deepSplit = F
)
}
\arguments{
\item{x}{an STList with normalized expression data.}

\item{weights}{a double \link{0-1} indicating the weight to be applied to spatial distances.
Defaults to w=0.025.}

\item{pcs}{the number of principal components (PCs) to retain.}

\item{vperc}{the minimum percentage of explained variance explained by the PCs to
decide how many of them to retain.}

\item{method}{the linkage method applied to hierarchical clustering. It is passed
to \code{hclust} and defaults to 'complete'.}

\item{ks}{the range of k values to assess. Defaults to \code{dtc}, meaning \code{cutreeDynamic}
is applied.}

\item{topgenes}{the number of high spot-to-spot standard deviation to retain before PCA.}

\item{deepSplit, }{a logical or integer \link{1-4}, to be passed to \code{cutreeDynamic} and
control cluster resolution.}
}
\value{
x, the STList with cluster assignments.
}
\description{
Perform unsupervised spatially-informed clustering on the spots of an
spatial array.
}
\details{
The function takes an STList and performs PCA on the gene expression data, then
calculates euclidean distances of the PCs and spatial distances. Weighted averages
between the two distances are calculated to perform hierarchical clustering. The
user can define how much weight the spatial distance should have (values
between 0.05 to 0.25 work reasonably well). The method aims to detect compartments
within a tissue.
}
\examples{
# In this example, melanoma is an STList.
# Using Dynamic Tree Cuts:
melanoma <- cluster_STspot(melanoma, ks='dtc', weights=0.1)

# Using a range of ks:
# melanoma <- cluster_STspot(melanoma, ks=c(2:6), weights=0.1)

}
