% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STclust.R
\name{STclust}
\alias{STclust}
\title{STclust: Detect clusters on ST arrays}
\usage{
STclust(
  x = NULL,
  weights = 0.025,
  dist = "euclidean",
  linkage = "ward.D",
  ks = "dtc",
  topgenes = 2000,
  deepSplit = F
)
}
\arguments{
\item{x}{an STList with normalized expression data.}

\item{weights}{a double (0-1) indicating the weight to be applied to spatial distances.
Defaults to w=0.025.}

\item{dist, }{the distrance measurement to be used. Defaults to 'euclidean'. Other
options as provided by \code{wordspace::dist.matrix}.}

\item{linkage}{the linkage method applied to hierarchical clustering. Passed to
\code{hclust} and defaults to 'ward.D'.}

\item{ks}{the range of k values to assess. Defaults to \code{dtc}, meaning \code{cutreeDynamic}
is applied.}

\item{topgenes}{the number of genes with highest spot-to-spot expression variation. The
variance is calculated via \code{Seurat::FindVariableFeatures}.}

\item{deepSplit, }{a logical or integer (1-4), to be passed to \code{cutreeDynamic} and
control cluster resolution.}
}
\value{
x, the STList with cluster assignments.
}
\description{
Perform unsupervised spatially-informed clustering on the spots of an
spatial array.
}
\details{
The function takes an STList and calculates euclidean distances between spots based on the x,y
spatial locations and expression of the top variable genes (\code{Seurat::FindVariableFeatures}).
Weighted averages between the two distances are calculated to perform hierarchical clustering. The
user can define how much weight the spatial distance should have (values
between 0.05 to 0.25 work reasonably well). The method aims to detect compartments
within a tissue.
}
\examples{
# In this example, melanoma is an STList.
# Using Dynamic Tree Cuts:
# melanoma <- STclust(melanoma, ks='dtc', weights=0.1)

# Using a range of ks:
# melanoma <- STclust(melanoma, ks=c(2:6), weights=0.1)

}
