% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdiff_core.R
\name{STdiff_compile_results}
\alias{STdiff_compile_results}
\title{STdiff_compile_results: Compile non-spatial and spatial results}
\usage{
STdiff_compile_results(prep = NULL, spatial = NULL)
}
\arguments{
\item{prep}{list from STdiff_select_genes}

\item{spatial}{list from STdiff_fit_spatial containing spatial model results}
}
\value{
list of data frames with final differential expression results per sample
}
\description{
Merges non-spatial and spatial model results into final output
}
\keyword{internal}
