% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STenrich.R, R/STenrich_legacy.R
\name{STenrich}
\alias{STenrich}
\alias{STenrich_legacy}
\title{STenrich: Test for spatial enrichment of gene expression sets}
\usage{
STenrich(
  x,
  samples = NULL,
  gene_sets = NULL,
  score_type = "avg",
  reps = 1000,
  annot = NULL,
  domain = NULL,
  num_sds = 1,
  min_units = 20,
  min_genes = 5,
  pval_adj_method = "BH",
  seed = 12345,
  cores = NULL,
  verbose = TRUE
)

STenrich_legacy(
  x = NULL,
  samples = NULL,
  gene_sets = NULL,
  score_type = "avg",
  reps = 1000,
  annot = NULL,
  domain = NULL,
  num_sds = 1,
  min_units = 20,
  min_genes = 5,
  pval_adj_method = "BH",
  seed = 12345,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{an STlist with transformed gene expression}

\item{samples}{a vector with sample names or indexes to run analysis}

\item{gene_sets}{a named list of gene sets to test. The names of the list should
identify the gene sets to be tested}

\item{score_type}{Controls how gene set expression is calculated. The options are
the average expression among genes in a set ('avg'), or a GSEA score ('gsva'). The
default is 'avg'}

\item{reps}{the number of random samples to be extracted. Default is 1000 replicates}

\item{annot}{name of the annotation within \code{x@spatial_meta} containing the spot/cell
categories. Needs to be used in conjunction with \code{domain}}

\item{domain}{the domain to restrict the analysis. Must exist within the spot/cell
categories included in the selected annotation (i.e., \code{annot})}

\item{num_sds}{the number of standard deviations to set the minimum gene set
expression threshold. Default is one (1) standard deviation}

\item{min_units}{Minimum number of spots with high expression of a pathway for
that gene set to be considered in the analysis. Defaults to 20 spots or cells}

\item{min_genes}{the minimum number of genes of a gene set present in the data set
for that gene set to be included. Default is 5 genes}

\item{pval_adj_method}{the method for multiple comparison adjustment of p-values.
Options are the same as that of \code{p.adjust}. Default is 'BH'}

\item{seed}{the seed number for the selection of random samples. Default is 12345}

\item{cores}{the number of cores used during parallelization. If NULL (default),
the number of cores is defined automatically}

\item{verbose}{logical, whether to print text to console}
}
\value{
a list of data frames with the results of the test

list of data frames with p-values and adjusted p-values per sample
}
\description{
Test for spatial enrichment of gene expression sets in spatial transcriptomics data

Original implementation of STenrich for reproducibility with previous results

Please use STenrich() for new analyses.
}
\details{
The function performs a randomization test to assess if the sum of
distances between cells/spots with high expression of a gene set is lower than
the sum of distances among randomly selected cells/spots. The cells/spots are
considered as having high gene set expression if the average expression of genes in a
set is higher than the average expression plus \code{num_sds} times the standard deviation.
Control over the size of regions with high expression is provided by setting the
minimum number of cells/spots (\code{min_units}). This method is a modification of
the method devised by Hunter et al. 2021 (zebrafish melanoma study).
}
\note{
This is the original implementation. Use STenrich() for new work.
}
\examples{
\donttest{
# Using included melanoma example (Thrane et al.)
# Download example data set from spatialGE_Data
thrane_tmp = tempdir()
unlink(thrane_tmp, recursive=TRUE)
dir.create(thrane_tmp)
lk='https://github.com/FridleyLab/spatialGE_Data/raw/refs/heads/main/melanoma_thrane.zip?download='
tryCatch({
  download.file(lk, destfile=paste0(thrane_tmp, '/', 'melanoma_thrane.zip'), mode='wb')
  unzip(file=paste0(thrane_tmp, '/melanoma_thrane.zip'), exdir=thrane_tmp)
  count_files <- list.files(paste0(thrane_tmp, '/melanoma_thrane'), full.names=TRUE, pattern='counts')
  coord_files <- list.files(paste0(thrane_tmp, '/melanoma_thrane'), full.names=TRUE, pattern='mapping')
  clin_file <- list.files(paste0(thrane_tmp, '/melanoma_thrane'), full.names=TRUE, pattern='clinical')
  st_obj <- STlist(rnacounts=count_files, spotcoords=coord_files, samples=clin_file)
  st_obj <- transform_data(st_obj)

  # Define gene sets (using first 10 genes from the data as example)
  genes <- rownames(st_obj@tr_counts[[1]])
  gene_sets <- list(
    GS1 = genes[1:5],
    GS2 = genes[6:10]
  )

  # Run STenrich
  res <- STenrich(st_obj, gene_sets = gene_sets, samples = 1, reps = 100)
}
}

}
\seealso{
STenrich_legacy for reproducibility with original implementation
}
