% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient.R
\name{STgradient}
\alias{STgradient}
\title{STgradient}
\usage{
STgradient(
  x = NULL,
  samples = NULL,
  topgenes = 2000,
  annot = NULL,
  ref = NULL,
  exclude = NULL,
  out_rm = F,
  limit = NULL,
  distsumm = "min",
  min_nb = 3,
  robust = T,
  cores = NULL
)
}
\arguments{
\item{x}{an STlist with transformed gene expression}

\item{cores}{the number of cores used during parallelization. If NULL (default),
the number of cores is defined automatically}
}
\value{
a list of data frames with the results of the test
}
\description{
STgradient
}
\details{
XXXXXXX
XXXXXXX
}
