% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_cleanup}
\alias{STgradient_cleanup}
\title{STgradient_cleanup: Remove empty samples and report timing}
\usage{
STgradient_cleanup(results_ls, zero_t, verbose)
}
\arguments{
\item{results_ls}{list of results data.frames}

\item{zero_t}{start time}

\item{verbose}{verbosity level}
}
\value{
cleaned list of results
}
\description{
Clean results by removing samples with no output and report completion time
}
\keyword{internal}
