% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_extract_expression}
\alias{STgradient_extract_expression}
\title{STgradient_extract_expression: Extract expression and join with distances}
\usage{
STgradient_extract_expression(x, sample_name, vargenes, dists_summ)
}
\arguments{
\item{x}{an STlist object}

\item{sample_name}{name of sample to process}

\item{vargenes}{vector of variable gene names}

\item{dists_summ}{data.frame with barcode and distance}
}
\value{
data.frame with expression data and distances
}
\description{
Extract transformed expression data for variable genes and join with distance data
}
\keyword{internal}
