% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_identify_variable_genes}
\alias{STgradient_identify_variable_genes}
\title{STgradient_identify_variable_genes: Identify top variable genes}
\usage{
STgradient_identify_variable_genes(x, sample_name, dists_summ, topgenes)
}
\arguments{
\item{x}{an STlist object}

\item{sample_name}{name of sample to process}

\item{dists_summ}{data.frame with distance data}

\item{topgenes}{number of top genes to select}
}
\value{
vector of gene names
}
\description{
Identify top genes by variance using VST method
}
\keyword{internal}
