% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_prepare_distances}
\alias{STgradient_prepare_distances}
\title{STgradient_prepare_distances: Calculate Euclidean distances}
\usage{
STgradient_prepare_distances(x, sample_name, annot, ref, exclude)
}
\arguments{
\item{x}{an STlist object}

\item{sample_name}{name of sample to process}

\item{annot}{annotation column name}

\item{ref}{reference cluster}

\item{exclude}{optional exclude cluster}
}
\value{
list with distance matrix and categorized spots
}
\description{
Calculate Euclidean distance matrix and categorize spots by cluster
}
\keyword{internal}
