% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_summarize_distances}
\alias{STgradient_summarize_distances}
\title{STgradient_summarize_distances: Summarize distances from reference}
\usage{
STgradient_summarize_distances(
  dist_tmp,
  nonref_tmp,
  ref_tmp,
  nbs_keep,
  distsumm,
  limit
)
}
\arguments{
\item{dist_tmp}{distance matrix}

\item{nonref_tmp}{non-reference barcodes}

\item{ref_tmp}{reference barcodes}

\item{nbs_keep}{reference barcodes to keep}

\item{distsumm}{distance summary metric ('min' or 'avg')}

\item{limit}{optional distance limit}
}
\value{
data frame with barcode and dist2ref
}
\description{
Calculate min or average distance from reference for non-reference spots
}
\keyword{internal}
