% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STgradient_helpers.R
\name{STgradient_validate_input}
\alias{STgradient_validate_input}
\title{STgradient_validate_input: Validate STgradient inputs}
\usage{
STgradient_validate_input(
  x,
  samples,
  topgenes,
  annot,
  ref,
  exclude,
  min_nb,
  nb_dist_thr,
  cores,
  verbose
)
}
\arguments{
\item{x}{STlist object}

\item{samples}{samples to process}

\item{topgenes}{number of top genes}

\item{annot}{annotation column}

\item{ref}{reference cluster}

\item{exclude}{optional exclude cluster}

\item{min_nb}{minimum neighbors}

\item{nb_dist_thr}{neighborhood distance threshold}

\item{cores}{number of cores}

\item{verbose}{verbosity}
}
\value{
validated parameters list
}
\description{
Validate inputs for STgradient function
}
\keyword{internal}
