% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SThet.R
\name{SThet}
\alias{SThet}
\title{SThet: Compute gene-wise global spatial autocorrelation statistics}
\usage{
SThet(x = NULL, genes = NULL, samples = NULL, method = "moran")
}
\arguments{
\item{x}{an STlist}

\item{genes}{a vector of gene names to compute statistics}

\item{samples}{the samples to compute statistics}

\item{method}{The spatial statistic(s) to estimate. Default is 'moran',}

\item{dist_thr}{the number of euclidean distance units to consider a cell/spot a neighbor.
but a vector with any combination of 'moran' or 'geary' can be passed.}
}
\description{
Computes Moran's I and/or Geary's C for a set of genes
}
\details{
The function computes global spatial autocorrelation statistics for the
requested genes. The spatial weights are defined by the minimum distance between
cells/spots. Spots within the two times the minimum distance are defined as neighbors and
a binary matrix is used as spatial weights for computation
}
