% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SThet_invdist_test.R
\name{SThet_invdist_test}
\alias{SThet_invdist_test}
\title{SThet: Computes global spatial autocorrelation statistics on gene expression}
\usage{
SThet_invdist_test(
  x = NULL,
  genes = NULL,
  samples = NULL,
  method = "moran",
  k = NULL,
  overwrite = T,
  cores = NULL
)
}
\arguments{
\item{x}{an STlist}

\item{genes}{a vector of gene names to compute statistics}

\item{samples}{the samples to compute statistics}

\item{method}{The spatial statistic(s) to estimate. It can be set to 'moran',
'geary' or both. Default is 'moran'}

\item{k}{the number of neighbors to estimate weights. By default NULL, meaning that
spatial weights will be estimated from Euclidean distances. If an positive integer is
entered, then the faster k nearest-neighbors approach is used. Please keep in mind
that estimates are not as accurate as when using the default distance-based method.}

\item{overwrite}{logical indicating if previous statistics should be overwritten.
Default to FALSE (do not overwrite)}
}
\value{
an STlist containing spatial statistics
}
\description{
Computes the global spatial autocorrelation statistics Moran's I and/or
Geary's C for a set of genes
}
\details{
The function computes global spatial autocorrelation statistics (Moran's I and/or
Geary's C) for the requested genes and samples. Then computation uses the
package \code{spdep}. The calculated statistics are stored in the STlist, which can
be accessed with the \code{get_gene_meta} function. For visual comparative analysis,
the function \code{compare_SThet} can be used afterwards.
}
