% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{class}
\name{STList-class}
\alias{STList-class}
\title{show}
\arguments{
\item{object, }{an STList object to show summary from.}
}
\description{
Prints summary of STList
}
\details{
This function takes an STList and prints a the number of spatial arrays in that
object.
}
