% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\docType{class}
\name{STlist-class}
\alias{STlist-class}
\title{Definition of an STlist object class.}
\description{
Definition of an STlist object class.
}
\section{Slots}{

\describe{
\item{\code{counts}}{per spot RNA counts}

\item{\code{spatial_meta}}{per spot x,y coordinates}

\item{\code{gene_meta}}{per gene statistics (e.g., average expression, variance, Moran's I)}

\item{\code{sample_meta}}{dataframe with metadata per sample}

\item{\code{tr_counts}}{transfromed per spot counts}

\item{\code{gene_krige}}{results from kriging on gene expression}

\item{\code{misc}}{Parameters and images from ST data}
}}

