% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STList.R
\name{STlist}
\alias{STlist}
\title{STlist: Creation of STlist objects for spatial transcriptomics analysis}
\usage{
STlist(
  rnacounts = NULL,
  spotcoords = NULL,
  samples = NULL,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{rnacounts}{the count data which can be provided in various formats (see \code{\link{STList_legacy}} for detailed legacy format descriptions, or new modular documentation).}

\item{spotcoords}{the cell/spot coordinates. Not required if inputs are Visium or Xenium.}

\item{samples}{the sample names/IDs and (optionally) metadata.}

\item{cores}{integer indicating the number of cores to use during parallelization.}

\item{verbose}{logical, whether to print text to console.}
}
\value{
an STlist object containing the counts and coordinates, and optionally sample metadata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates an STlist object from one or multiple spatial transcriptomic samples.
}
\details{
Objects of the S4 class STlist are the starting point of analyses in \strong{\code{spatialGE}}.

\strong{Note:} This is the new modular implementation of \code{STList} created in version 2.0.0.
For reproducibility, the previous version is available as \code{STList_legacy}.

The STlist contains data from one or multiple samples (i.e., tissue slices), and
results from most \code{spatialGE}'s functions are stored within the object.
}
\seealso{
\code{\link{STList_legacy}}
}
