% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STList.R
\name{STList}
\alias{STList}
\title{STList: Creation of STList objects}
\usage{
STList(
  rnacounts = NULL,
  spotcoords = NULL,
  samples = NULL,
  gmx_pkc = NULL,
  gmx_slide_col = NULL,
  gmx_roi_col = NULL,
  gmx_x_col = NULL,
  gmx_y_col = NULL,
  gmx_meta_cols = NULL
)
}
\arguments{
\item{rnacounts}{The count data which can be provided in one of these formats:
\itemize{
\item File paths to comma- or tab-delimited files containing raw RNA counts, one
for each spatial array. The first column contains gene names. Subsequent columns
contain data for each spot. Duplicate gene names will be appended a suffix ('_d**').
Requires \code{spotcoords} and \code{samples}.
\item File paths to Visium output folders (one per spatial array). The folder should have the
structure resulting from \verb{spaceranger count} and contain the \code{.h5} files and \code{spatial} folder.
Requires \code{samples}.
\item File path to \code{.dcc} files from GeoMx output. Requires \code{samples}.
\item One named list of data frames with raw RNA counts (one data frame per spatial array).
Requires \code{spotcoords} and \code{samples}.
}}

\item{spotcoords}{The spot coordinates, which are required if inputs are not Visium outputs.
The user can provide one of these formats:
\itemize{
\item File paths to comma- or tab-delimited files containing spot coordinates, one
for each spatial array. The files must contain three columns: spot IDs, y positions, and
x positions. The spot IDs must match the column names for each spot (column) in the RNA count files.
\item One named list of data frames with spot coordinates. The list names must match list
names of the RNA counts list.
}}

\item{samples}{The metadata associated to each spatial array. This file can also include
system paths to RNA count and spot coordinate files, bypassing the need to specify \code{rnacounts}
and \code{spotcoords}. One of the following options should be entered to create an STList:
\itemize{
\item A vector with sample names, which will be used to partial match RNA count and
spot coordinates file paths. A sample name must not match two file paths.
\item A path to  the file containing metadata for each spatial array. This file is a comma-
or tab-separated file with one sample per row and sample names in the first column.
Paths to RNA count and coordinate files can be in the second and third column respectively,
and omitting the \code{rnacounts} and \code{spotcoords} arguments. If Visium, only the second
column with paths to output folders is expected. Subsequent columns can contain
variables associated to each spatial array.
Note: For GeoMx, the metadata file contains one row per ROI. This information is later summarized
to one row per slide by this function.
}}

\item{gmx_pkc, }{the file path to the \code{.pkc} (for GeoMx input)}

\item{gmx_slide_col, }{the name of the column in the metadata table containing the slide names (for GeoMx input)}

\item{gmx_roi_col, }{the name of the column in the metadata table containing the ROI IDs, matching IDs in the DCC files (for GeoMx input)}

\item{gmx_x_col, }{the name of the column in the metadata table containing the x coordinates (for GeoMx input)}

\item{gmx_y_col, }{the name of the column in the metadata table containing the y coordinates (for GeoMx input)}

\item{gmx_meta_cols, }{a vector with column names in the metadata table containing clinical data (for GeoMx input)}
}
\value{
x, the STList object containing the counts and coordinates, and optionally
the sample metadata.
}
\description{
Creates an STList object from one or several spatial transcriptomic samples.
}
\details{
Objects of the S4 class STList can be created from two sources:
\itemize{
\item Raw RNA counts (genes in rows, spots in columns) and spot coordinates (spots
in rows, with three columns: spot IDs, y position, and x position).
\item Visium output folders from Space Ranger. The folder should have the
structure resulting from \verb{spaceranger count}.
}
Optionally, the user can input a system path to a table containing data associated with
each spatial array (e.g., clinical variables). This sample metadata file should contain
sample IDs in the first column matching partially the file names of the
count/coordinate file paths or Visium directories.

The function will read data in parallel if unix system is available (Windows users
will experience longer times depending on the number of samples).
}
\examples{
# Using included melanoma example (Thrane et al.)
# data_files <- system.file("extdata", package="spatialGEdev")
# count_files <- grep("genes", data_files, value=T)
# coord_files <- grep("mapping", data_files, value=T)
# clin_file <- grep("clinical", data_files, value=T)
# melanoma <- STList(rnacounts=count_files, spotcoords=coord_files, samples=clin_file)
# melanoma
}
