% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_STlist.R
\name{summarize_STlist}
\alias{summarize_STlist}
\title{summarize_STlist: Generates a data frame with summary statistics}
\usage{
summarize_STlist(x = NULL)
}
\arguments{
\item{x}{an STlist}
}
\value{
a data frame
}
\description{
Produces a data frame with counts per gene and counts per ROI/spot/cell
}
\details{
The function creates a table with counts per gene and counts per region of interest (ROI),
spot, or cell in the samples stored in the STlist
}
\examples{
# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- list.files(system.file("extdata", package="spatialGE"), recursive=T, full.names=T)
count_files <- grep("counts", data_files, value=T)
coord_files <- grep("mapping", data_files, value=T)
clin_file <- grep("thrane_clinical", data_files, value=T)
melanoma <- STlist(rnacounts=count_files, spotcoords=coord_files, samples=clin_file)
summarize_STlist(melanoma)

}
