% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tissue_names}
\alias{tissue_names}
\title{tissue_names: Prints the names of the tissue samples in the STlist}
\usage{
tissue_names(x)
}
\arguments{
\item{x}{an STList object}
}
\value{
a character vector with the sample names in the STlist object
}
\description{
returns a character vector with the names of tissue samples in the
STlist.
}
