% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTransform.R
\name{voom_norm}
\alias{voom_norm}
\title{voom_norm: voom transformation of ST arrays}
\usage{
voom_norm(x = NULL)
}
\arguments{
\item{x}{an STList with raw count matrices.}
}
\value{
x an updated STList with transformed counts.
}
\description{
Applies limma-voom transformation to spatial transcriptomics arrays data.
}
\details{
This function takes an STList and returns roughly Gaussian count data matrices
in two steps. In the first step, (edgeR) normalization factors are used to scale
the counts from each library/spot. In the second step, limma-voom transformation
is applied. The resulting normalized count matrix is stored in the voom_counts
slot of the STList. The function also calculates gene-wise standard deviation from
the transformed counts and stores them in the gene_stdev slot.

The function works on parallel using "forking" (not in Windows OS).
}
\examples{
# In this example, melanoma is an STList.
# melanoma <- voom_norm(melanoma)

}
